### 2021-07-23 #43 @weiji14
PyGMT:  A Python interface for the Generic Mapping Tools
Submitting Author: Wei Ji Leong (@weiji14)
Package Name: PyGMT
One-Line Description of Package: A Python interface for the Generic Mapping Tools
Repository Link: https://github.com/GenericMappingTools/pygmt
Version submitted:
Archive: [Zenodo Archive](https://zenodo.org/record/6702566)
Version accepted: V 0.7.0

---

    PyGMT is a library for processing geospatial and geophysical data and making publication quality maps and figures. It provides a Pythonic interface for the Generic Mapping Tools (GMT), a command-line program widely used in the Earth Sciences.

  - Data munging :gear:: Several data processing modules are available to work on tabular and grid data (https://www.pygmt.org/v0.4.0/api/index.html#data-processing) including functions like `blockmedian`, `surface`, `grdtrack`, etc
  - Data visualization :world_map:: Plotting publication quality maps and figures is a strong point of PyGMT (https://www.pygmt.org/v0.4.0/api/index.html#plotting). There is a comprehensive set of tools for plotting 2D and 3D data, both geospatial and non-geospatial. It is also possible to add standard map elements like a `colorbar` or `legend`, and make multi-panel `subplot` figures.
  - Geospatial :globe_with_meridians:: PyGMT is inherently a geospatial package that can handle both vector and raster data. Several map projections are supported (https://www.pygmt.org/v0.4.0/projections/index.html), and there is support for plotting shapefiles and GeoTiff/NetCDF files.

  - The primary target audience of PyGMT is the geoscience community, including anybody working on fields like Earth Observation, Geophysics, Oceanography, Seismology, Planetary Sciences, etc. The package offers a way for users to perform general geoprocessing tasks and create high quality illustrations of their results for posters or publications.

  Yes, there are several geospatial Python packages such as:

  - [`cartopy`](https://github.com/SciTools/cartopy) for plotting vector data ([`matplotlib`](https://github.com/matplotlib/matplotlib) based)
  - [`geopandas`](https://geopandas.org/index.html) for working with vector data
  - [`rasterio`](https://github.com/mapbox/rasterio), [`rioxarray`](https://github.com/corteva/rioxarray) and [`xarray-spatial`](https://github.com/makepath/xarray-spatial) for working with raster data

  These tools are typically focused on one thing only (e.g. plotting maps, vector data, raster data). PyGMT allows users to **mix** vector and raster data easily, so that users can:
    - **Produce** a map with points/lines/polygons and raster grids in the same plot, e.g. plot weather station points on top of a raster Digital Elevation Model, while including map elements (e.g. scalebar, compass arrow, [colorbar](https://www.pygmt.org/v0.4.0/api/generated/pygmt.Figure.colorbar.html), [legend](https://www.pygmt.org/v0.4.0/api/generated/pygmt.Figure.legend.html), [inset](https://www.pygmt.org/v0.4.0/api/generated/pygmt.Figure.inset.html) overview maps etc)
    - **Extract** a series of elevation values along a transect line (vector) from a NetCDF/GeoTiff grid (raster) using [`grdtrack`](https://www.pygmt.org/v0.4.0/api/generated/pygmt.grdtrack.html)
    - **Convert** a table of xyz points (vector) to a 2D grid (raster) using [`surface`](https://www.pygmt.org/v0.4.0/api/generated/pygmt.surface.html) via a spline-based interpolation method.

  PyGMT integrates with the [PyData](https://pydata.org) ecosystem. It allows users to process and plot data stored in [NumPy](https://numpy.org) arrays, [Pandas](https://pandas.pydata.org) DataFrames, [Xarray](https://xarray.pydata.org/en/stable) DataArray/Datasets, [GeoPandas](https://geopandas.readthedocs.io/en/latest/) GeoDataFrames, as well as from standard file formats like CSV and GeoTiff/NetCDF files. There are also plans to integrate with other scientific libraries like ObsPy (https://github.com/GenericMappingTools/pygmt/issues/967) in the future.

  However, PyGMT can only produce **static** plots unlike packages like [Geoviews](https://geoviews.org) which allow for interactive plotting (e.g. panning and zooming). It is also non-trivial to scale out data processing tasks and/or plotting of big datasets (>10GB) that don't fit in RAM to multiple processors/clusters as with libraries like [`dask-geopandas`](https://github.com/geopandas/dask-geopandas) because of limitations in the [GMT](https://www.generic-mapping-tools.org/) C package that PyGMT wraps around.

    First off, thanks for reading this! We're keen to get some independent reviewers to have a look at the PyGMT package, sort of as a first step before going for a software publication at JOSS/G3 (the team is still deciding). Linking to original discussion at https://github.com/GenericMappingTools/pygmt/issues/677#issuecomment-880287165, cc @leouieda @seisman @meghanrjones.
