### 2021-07-23 8566bfc
Add how to use different colormaps for subplots
