### 2021-07-24 b97d039
Apply suggestions from code review
Co-authored-by: Will Schlitzer <schlitzer90@gmail.com>
