### 2021-07-24 fdbde07
Apply suggestions from code review
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
