### 2021-07-24 6b1e49d
add pytest fixture
