+++ pygmt/tests/test_xyz2grd.py
@@ -1,9 +1,17 @@
 """
 Tests for xyz2grd.
 """
+import pytest
 import xarray as xr
+from pygmt.datasets import load_sample_bathymetry
 from pygmt import xyz2grd
 
+@pytest.fixture(scope="module", name="ship_data")
+def fixture_ship_data():
+    """
+    Load the data from the sample bathymetry dataset.
+    """
+    return load_sample_bathymetry()
 
 def test_xyz2grd_input_file():
     """
