### 2021-07-24 63e2332
reorganize import in test_xyz2grd.py
