+++ pygmt/src/xyz2grd.py
@@ -45,36 +45,15 @@ def xyz2grd(table, **kwargs):
 
     Parameters
     ----------
-    table : ascii file (xyz tables)
-    The file name of the input xyz file, with the extension: file.xyz
+    table : str or {table-like}
+        Pass in either a file name to an ASCII data table, a 1D/2D
+        {table-classes}.
 
-    {G}: str or None
-    The name of the output netCDF file with extension .nc to store the grid
-    in. If non is given, it will be the same name of the input file, just with
-    different extensions.
-
-    I : str
-    xinc[unit][=|+][/yinc[unit][=|+]]
-    x_inc [and optionally y_inc] is the grid spacing. Optionally, append a suffix modifier.
-    Geographical (degrees) coordinates: Append m to indicate arc minutes or s to indicate arc
-    seconds. If one of the units e, f, k, M, n or u is appended instead, the increment is
-    assumed to be given in meter, foot, km, Mile, nautical mile or US survey foot,
-    respectively, and will be converted to the equivalent degrees longitude at the middle
-    latitude of the region (the conversion depends on PROJ_ELLIPSOID). If y_inc is given but
-    set to 0 it will be reset equal to x_inc; otherwise it will be converted to degrees
-    latitude. All coordinates: If = is appended then the corresponding max x (east) or y
-    (north) may be slightly adjusted to fit exactly the given increment [by default the
-    increment may be adjusted slightly to fit the given domain]. Finally, instead of giving an
-    increment you may specify the number of nodes desired by appending + to the supplied
-    integer argument; the increment is then recalculated from the number of nodes and
-    the domain. The resulting increment value depends on whether you have selected a gri
-    dline-registered or pixel-registered grid; see GMT File Formats for details. Note: if
-    -Rgrdfile is used then the grid spacing has already been initialized; use -I to override
-    the values.
-
-    {R}: str
-    [unit]xmin/xmax/ymin/ymax[r] (more ...)
-    Specify the region of interest.
+    outgrid : str or None
+        The name of the output netCDF file with extension .nc to store the grid
+        in.
+    {I}
+    {R}
 
     Optional Arguments
     ------------------
