### 2021-07-24 b150564
remove "Optional Arguments" sub-header
