+++ examples/gallery/embellishments/colorbars_multiple.py
@@ -2,7 +2,7 @@
 Multiple colormaps
 ------------------
 This gallery example shows how to create multiple colormaps for different subplots. To
-better understand how GMT modern mode maintain several levels of colormaps,
+better understand how GMT modern mode maintains several levels of colormaps,
 please refer to :gmt-docs:`cookbook/features.html#gmt-modern-mode-hierarchical-levels`
 for details.
 """
