### 2021-07-26 dc5cf3e
Add gallery example for grdclip
Since noone worked on this issue during the scipy-sprint here's a gallery example adressing #1384.
