### 2021-07-27 751f832
Update pygmt/datasets/samples.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
