+++ pygmt/src/xyz2grd.py
@@ -32,6 +32,8 @@ from pygmt.helpers import (
 @kwargs_to_strings(R="sequence")
 def xyz2grd(table, **kwargs):
     """
+    Create a grid file from table data.
+
     xyz2grd reads one or more z or xyz tables and creates a binary grid file.
     xyz2grd will report if some of the nodes are not filled in with data. Such
     unconstrained nodes are set to a value specified by the user [Default is
