+++ pygmt/__init__.py
@@ -43,7 +43,6 @@ from pygmt.src import (
     grdtrack,
     info,
     makecpt,
-    sphdistance,
     surface,
     which,
     x2sys_cross,
