### 2021-07-28 15f9325
update test_hotspots() in test_datasets_samples.py
