### 2021-07-29 d5c9037
Update examples/gallery/images/grdclip.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
