+++ examples/gallery/images/grdclip.py
@@ -1,6 +1,6 @@
 """
-Clip grid values
-----------------
+Clipping grid values
+--------------------
 The :meth:`pygmt.grdclip` method allows to clip defined ranges of grid values.
 In the example shown below we set all elevation values (grid points) smaller
 than 0 m (in general the bathymetric part of the grid) to a common value of
