### 2021-07-29 99822bd
Add 1 arc minute global relief data to cache
As discussed in #1396 this PR adds the 1 arc minute global relief data to cache.
