+++ pygmt/helpers/testing.py
@@ -157,6 +157,8 @@ def download_test_data():
         "@earth_relief_10m_p",
         "@earth_relief_05m_p",
         "@earth_relief_05m_g",
+        "@earth_relief_01m_p",
+        "@earth_relief_01m_g",
         # List of tiles of 03s srtm data.
         # Names like @N35E135.earth_relief_03s_g.nc is for internal use only.
         # The naming scheme may change. DO NOT USE IT IN YOUR SCRIPTS.
