### 2021-07-29 b48cad9
Add grdhisteq processing function
