+++ doc/api/index.rst
@@ -92,6 +92,7 @@ Operations on grids:
     grdcut
     grdfill
     grdfilter
+    grdhisteq
     grdlandmask
     grdgradient
     grdtrack
