### 2021-07-29 f2e91ea
Add function wrapping grdconvert
