### 2021-07-31 afe8719
Temporarily viewing verbose output from pip install step
