+++ .github/workflows/ci_tests_dev.yaml
@@ -6,7 +6,7 @@ on:
   # push:
   #   branches: [ master ]
   pull_request:
-    types: [ready_for_review]
+    types: [synchronize, ready_for_review]
     paths-ignore:
       - 'doc/**'
       - 'examples/**'
@@ -85,7 +85,7 @@ jobs:
         run: |
           conda install ninja cmake libblas libcblas liblapack fftw gdal geopandas \
                         ghostscript libnetcdf hdf5 zlib curl pcre make dvc
-          pip install --pre numpy pandas xarray netCDF4 packaging \
+          pip install --verbose --pre numpy pandas xarray netCDF4 packaging \
                             ipython pytest-cov pytest-mpl pytest>=6.0 sphinx-gallery
 
       # Build and install latest GMT from GitHub
