### 2021-07-31 0a8ab3d
List installed dependencies from conda and pip
