+++ .github/workflows/ci_tests_dev.yaml
@@ -86,7 +86,8 @@ jobs:
           conda install ninja cmake libblas libcblas liblapack fftw gdal geopandas \
                         ghostscript libnetcdf hdf5 zlib curl pcre make dvc
           pip install --verbose --pre numpy pandas xarray netCDF4 packaging \
-                            ipython pytest-cov pytest-mpl pytest>=6.0 sphinx-gallery
+                            ipython pytest-cov pytest-mpl pytest>=6.0 sphinx-gallery \
+                            toml
           conda list
           pip list
 
