### 2021-07-31 34c3d1e
update docstring for xyz2grd.py
