+++ pygmt/src/xyz2grd.py
@@ -57,84 +57,93 @@ def xyz2grd(table, **kwargs):
     {I}
     {R}
     -A[f|l|m|n|r|s|u|z]
-    By default we will calculate mean values if multiple entries fall on the same node.
-    Use -A to change this behavior, except it is ignored if -Z is given. Append f or s
-    to simply keep the first or last data point that was assigned to each node. Append
-    l or u to find the lowest (minimum) or upper (maximum) value at each node, respectively.
-    Append m or r to compute mean or RMS value at each node, respectively. Append n to
-    simply count the number of data points that were assigned to each node (this only
-    requires two input columns x and y as z is not consulted). Append z to sum multiple
-    values that belong to the same node.
+    By default we will calculate mean values if multiple entries fall on
+    the same node. Use **-A** to change this behavior, except it is
+    ignored if **-Z** is given. Append **f** or **s** to simply keep the
+    first or last data point that was assigned to each node. Append
+    **l** or **u** or **d** to find the lowest (minimum) or upper (maximum)
+    value or the difference between the maximum and miminum value at each
+    node, respectively. Append **m** or **r** or **S** to compute mean or
+    RMS value or standard deviation at each node, respectively. Append **n**
+    to simply count the number of data points that were assigned to each node
+    (this only requires two input columns *x* and *y* as *z* is not consulted).
+    Append **z** to sum multiple values that belong to the same node.
 
     -Dxname/yname/zname/scale/offset/invalid/title/remark
-    Give values for xname, yname, zname (give the names of those variables and in square
-    bracket their units, e.g., “distance [km]”), scale (to multiply grid values after read
-    [normally 1]), offset (to add to grid after scaling [normally 0]), invalid (a value to
-    represent missing data [NaN]), title (anything you like), and remark (anything you like).
-    To leave some of these values untouched, leave field blank. Empty fields in the end may
-    be skipped. Alternatively, to allow “/” to be part of one of the values, use any
-    non-alphanumeric character (and not the equal sign) as separator by both starting and
-    ending with it. For example: -D:xname:yname:zname:scale:offset:invalid:title:remark:
-    Use quotes to group texts with more than one word. Note that for geographic grids (-fg)
-    xname and yname are set automatically.
+    Give values for xname, yname, zname (give the names of those variables and
+    in square bracket their units, e.g., “distance [km]”), scale (to multiply
+    grid values after read [normally 1]), offset (to add to grid after
+    scaling [normally 0]), invalid (a value to represent missing data [NaN]),
+    title (anything you like), and remark (anything you like). To leave some
+    of these values untouched, leave field blank. Empty fields in the end may
+    be skipped. Alternatively, to allow “/” to be part of one of the values,
+    use any non-alphanumeric character (and not the equal sign) as separator
+    by both starting and ending with it.
 
     -S[zfile]
     Swap the byte-order of the input only. No grid file is produced.
-    You must also supply the -Z option. The output is written to zfile (or stdout if not supplied).
-
-    -V[level] (more ...)
-    Select verbosity level [c].
+    You must also supply the -Z option. The output is written to zfile
+    (or stdout if not supplied).
+    {V}
 
     -Z[flags]
-    Read a 1-column ASCII [or binary] table. This assumes that all the nodes are present
-    and sorted according to specified ordering convention contained in flags. If incoming
-    data represents rows, make flags start with T(op) if first row is y = ymax or B(ottom)
-    if first row is y = ymin. Then, append L or R to indicate that first element is at left
[…]
