### 2021-07-31 23f6343
add "@hotspots.txt" to testing.py
