### 2021-07-31 ba5c105
updates based on code review
