### 2021-07-31 57df81f
used 3 arc min grid instead of 1 arc min
