### 2021-08-01 4775c18
Update doc/contributing.md
Co-authored-by: Yao Jiayuan <coreman.seism@gmail.com>
