### 2021-08-01 7186530
Update pygmt/helpers/testing.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
