### 2021-08-01 a1401bd
Update examples/gallery/images/grdclip.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
