+++ examples/gallery/images/grdclip.py
@@ -22,6 +22,7 @@ fig.basemap(region=region, projection="M12c", frame=["f", '+t"original grid"'])
 fig.grdimage(grid=grid, cmap="oleron")
 
 # Shift plot origin of the second map by "width of the first map + 0.5 cm"
+# in x direction
 fig.shift_origin(xshift="w+0.5c")
 
 # Set all grid points < 0 m to a value of -2000 m.
