+++ pygmt/src/plot.py
@@ -209,8 +209,13 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
     extra_arrays = []
     if "S" in kwargs and kwargs["S"][0] in "vV" and direction is not None:
         extra_arrays.extend(direction)
-    elif "S" not in kwargs and data is not None and '_geometry_column_name' in data.__dict__ and (data.geom_type == 'Point').all():
-        kwargs["S"] = 's0.2c'
+    elif (
+        "S" not in kwargs
+        and data is not None
+        and "_geometry_column_name" in data.__dict__
+        and (data.geom_type == "Point").all()
+    ):
+        kwargs["S"] = "s0.2c"
     if "G" in kwargs and not isinstance(kwargs["G"], str):
         if kind != "vectors":
             raise GMTInvalidInput(
