### 2021-08-02 af36475
chcking that data is an object to be sure it has __dict__ (make sure it is not a string)
