### 2021-08-02 e1543e8
Update pygmt/src/plot.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
