### 2021-08-02 d60ee1b
not checking if data is not None allrady been verefide in data_kind function
