### 2021-08-02 4df73a0
updates based on code review
