+++ pygmt/helpers/testing.py
@@ -157,11 +157,11 @@ def download_test_data():
         "@earth_relief_10m_p",
         "@earth_relief_05m_p",
         "@earth_relief_05m_g",
-        "@earth_relief_03m_p",
         # List of tiles of 03s srtm data.
         # Names like @N35E135.earth_relief_03s_g.nc is for internal use only.
         # The naming scheme may change. DO NOT USE IT IN YOUR SCRIPTS.
         "@N35E135.earth_relief_03s_g.nc",
+        "@N00W090.earth_relief_03m_p.nc",
         # Other cache files
         "@fractures_06.txt",
         "@ridge.txt",
