### 2021-08-02 8a3f74d
Add binary alias
