+++ pygmt/src/wiggle.py
@@ -21,6 +21,7 @@ from pygmt.helpers import build_arg_string, fmt_docstring, kwargs_to_strings, us
     Z="scale",
     b="binary",
     c="panel",
+    d="nodata",
     i="columns",
     p="perspective",
 )
@@ -78,6 +79,7 @@ def wiggle(self, x=None, y=None, z=None, data=None, **kwargs):
     {XY}
     {b}
     {c}
+    {d}
     columns : str or 1d array
         Choose which columns are x, y, and z, respectively if input is provided
         via *data*. E.g. ``columns = [0, 1, 2]`` or ``columns = "0,1,2"`` if
