+++ pygmt/src/wiggle.py
@@ -22,6 +22,7 @@ from pygmt.helpers import build_arg_string, fmt_docstring, kwargs_to_strings, us
     b="binary",
     c="panel",
     d="nodata",
+    e="find",
     i="columns",
     p="perspective",
 )
@@ -80,6 +81,7 @@ def wiggle(self, x=None, y=None, z=None, data=None, **kwargs):
     {b}
     {c}
     {d}
+    {e}
     columns : str or 1d array
         Choose which columns are x, y, and z, respectively if input is provided
         via *data*. E.g. ``columns = [0, 1, 2]`` or ``columns = "0,1,2"`` if
