+++ pygmt/src/wiggle.py
@@ -24,6 +24,7 @@ from pygmt.helpers import build_arg_string, fmt_docstring, kwargs_to_strings, us
     d="nodata",
     e="find",
     g="gap",
+    h="header",
     i="columns",
     p="perspective",
 )
@@ -84,6 +85,7 @@ def wiggle(self, x=None, y=None, z=None, data=None, **kwargs):
     {d}
     {e}
     {g}
+    {h}
     columns : str or 1d array
         Choose which columns are x, y, and z, respectively if input is provided
         via *data*. E.g. ``columns = [0, 1, 2]`` or ``columns = "0,1,2"`` if
