+++ pygmt/src/blockm.py
@@ -72,6 +72,7 @@ def _blockm(block_method, table, outfile, x, y, z, **kwargs):
     a="aspatial",
     f="coltypes",
     i="incols",
+    o="outcols",
     r="registration",
 )
 @kwargs_to_strings(R="sequence")
@@ -112,6 +113,7 @@ def blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     {a}
     {i}
     {f}
+    {o}
     {r}
 
     Returns
@@ -137,6 +139,7 @@ def blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     a="aspatial",
     f="coltypes",
     i="incols",
+    o="outcols",
     r="registration",
 )
 @kwargs_to_strings(R="sequence")
@@ -177,6 +180,7 @@ def blockmedian(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     {a}
     {f}
     {i}
+    {o}
     {r}
 
     Returns
