+++ pygmt/src/blockm.py
@@ -75,6 +75,7 @@ def _blockm(block_method, table, outfile, x, y, z, **kwargs):
     o="outcols",
     r="registration",
     s="skiprows",
+    w="wrap",
 )
 @kwargs_to_strings(R="sequence")
 def blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
@@ -117,6 +118,7 @@ def blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     {o}
     {r}
     {s}
+    {w}
 
     Returns
     -------
@@ -144,6 +146,7 @@ def blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     o="outcols",
     r="registration",
     s="skiprows",
+    w="wrap",
 )
 @kwargs_to_strings(R="sequence")
 def blockmedian(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
@@ -186,6 +189,7 @@ def blockmedian(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs):
     {o}
     {r}
     {s}
+    {w}
 
     Returns
     -------
