+++ pygmt/helpers/decorators.py
@@ -163,6 +163,9 @@ COMMON_OPTIONS = {
             (:gmt-term:`PROJ_MEAN_RADIUS`), and the specification of latitude type
             (:gmt-term:`PROJ_AUX_LATITUDE`). Geodesic distance calculations is also
             controlled by method (:gmt-term:`PROJ_GEODESIC`).""",
+    "l": r"""
+        label : str
+        Add a legend entry for the symbol or line being plotted.""",
     "n": r"""
         interpolation : str
             [**b**\|\ **c**\|\ **l**\|\ **n**][**+a**][**+b**\ *BC*][**+c**][**+t**\ *threshold*].
