### 2021-08-03 93c5312
Update pygmt/src/plot.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
