### 2021-08-03 5daeb42
Update pygmt/src/plot.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
