+++ pygmt/src/plot.py
@@ -209,7 +209,7 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
     extra_arrays = []
     if "S" in kwargs and kwargs["S"][0] in "vV" and direction is not None:
         extra_arrays.extend(direction)
-    elif "S" not in kwargs and kind == "geojson" and (data.geom_type == "Point").all():
+    elif "S" not in kwargs and kind == "geojson" and data.geom_type.isin(["Point", "MultiPoint"]).all():
         kwargs["S"] = "p"
     if "G" in kwargs and not isinstance(kwargs["G"], str):
         if kind != "vectors":
