### 2021-08-03 2041146
adding the if statment to plot3d.py
