### 2021-08-03 9f69437
changing the style to s-.2c and u0.2c for plot and plot3d respectivly
