+++ doc/contributing.md
@@ -95,7 +95,7 @@ to make changes to our codebase. Every change made goes through a pull request, 
 our own, so that our
 [continuous integration](https://en.wikipedia.org/wiki/Continuous_integration)
 services have a chance to check that the code is up to standards and passes all
-our tests. This way, the *master* branch is always stable.
+our tests. This way, the *main* branch is always stable.
 
 #### General Guidelines for Making a Pull Request (PR):
 
@@ -170,7 +170,7 @@ It will make your life a lot easier!
 
 The repository includes a conda environment file `environment.yml` with the
 specification for all development requirements to build and test the project.
-See the [`environment.yml`](https://github.com/GenericMappingTools/pygmt/blob/master/environment.yml)
+See the [`environment.yml`](https://github.com/GenericMappingTools/pygmt/blob/main/environment.yml)
 file for the list of dependencies and the environment name (`pygmt`).
 Once you have forked and cloned the repository to your local machine, you can
 use this file to create an isolated environment on which you can work.
