### 2021-08-03 bc55332
Update pygmt/helpers/decorators.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
