### 2021-08-03 ae628c8
Apply suggestions from code review
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
