### 2021-08-04 6409d34
testing defualt behavior of plot and plot3d with geopandas Point and MultiPoint DataFrames
