### 2021-08-04 0fbc83a
Update pygmt/src/nearneighbor.py
Enhance nearneighbour documentation

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
