### 2021-08-04 4ae19cb
Add similar figure to that found in GMTs nearneighbor documentation
