### 2021-08-04 9fdffc0
Move description for -l common option to pygmt/helpers/decorators.py (#1408)
* Move label description to COMMON OPTIONS

* Add link to full -l docs

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
