### 2021-08-04 9636ac1
Apply suggestions from code review
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
