### 2021-08-04 27b4f54
Apply suggestions from code review
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
