### 2021-08-04 cbd508b
Add common option aliases to COMMON_OPTIONS in decorators.py (#1407)
* Add binary alias

* Add nodata alias

* Add find alias

* Add gap alias

* Add header alias

* Add outcols alias

* Add skiprows alias

* Add wrap alias

* Format

* Apply suggestions from code review

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
