### 2021-08-04 9cbda9e
Change references from "master" branch to "main" branch (#1360)
PyGMT is changing its default branch from "master" to "main".
This pull request changes the references to "master" in the CI
job files, maintenance.md, and the README.

* change workflow files to point to "main" branch instead of "master" branch
* change README to use "main" branch instead of "master" branch
* change maintenance.md use "main" branch instead of "master" branch
* update conf.py to use main instead of master
* update contributing.md to use main instead of master
* update external_resources.md to use main instead of m
[…]
