+++ .github/workflows/ci_tests.yaml
@@ -4,7 +4,7 @@ name: Tests
 
 on:
   push:
-    branches: [ master ]
+    branches: [ main ]
   pull_request:
     types: [opened, reopened, synchronize, ready_for_review]
     paths-ignore:
