+++ .github/workflows/ci_docs.yml
@@ -76,7 +76,7 @@ jobs:
 
       # Download cached remote files (artifacts) from GitHub
       - name: Download remote data from GitHub
-        uses: dawidd6/action-download-artifact@v2.14.0
+        uses: dawidd6/action-download-artifact@v2.14.1
         with:
           workflow: cache_data.yaml
           workflow_conclusion: success
