### 2021-08-05 ad1440b
Revert verbose output printing
