### 2021-08-05 1f9ab43
Add tomli as a dependency in GMT Dev Tests (#1401)
Adding tomli to fix the GMT Dev Tests breakages reported
in #1392. This is a temporary measure until `pytest-cov`
releases a new version > 2.12.1.
