+++ .github/workflows/ci_tests_dev.yaml
@@ -86,7 +86,8 @@ jobs:
           conda install ninja cmake libblas libcblas liblapack fftw gdal geopandas \
                         ghostscript libnetcdf hdf5 zlib curl pcre make dvc
           pip install --pre numpy pandas xarray netCDF4 packaging \
-                            ipython pytest-cov pytest-mpl pytest>=6.0 sphinx-gallery
+                            ipython pytest-cov pytest-mpl pytest>=6.0 sphinx-gallery \
+                            tomli
 
       # Build and install latest GMT from GitHub
       - name: Install GMT ${{ matrix.gmt_git_ref }} branch (Linux/macOS)
