### 2021-08-05 b9288ec
Add changelog entry for v0.4.1
