+++ README.rst
@@ -154,7 +154,7 @@ Developers". Feel free to cite our work in your research using the following Bib
 
 .. code-block::
 
-    @software{pygmt_2021_4978645,
+    @software{pygmt_2021_5162003,
       author       = {Uieda, Leonardo and
                       Tian, Dongdong and
                       Leong, Wei Ji and
@@ -169,12 +169,12 @@ Developers". Feel free to cite our work in your research using the following Bib
                       Ziebarth, Malte and
                       Wessel, Paul},
       title        = {{PyGMT: A Python interface for the Generic Mapping Tools}},
-      month        = jun,
+      month        = jul,
       year         = 2021,
       publisher    = {Zenodo},
-      version      = {v0.4.0},
-      doi          = {10.5281/zenodo.4978645},
-      url          = {https://doi.org/10.5281/zenodo.4978645}
+      version      = {v0.4.1},
+      doi          = {10.5281/zenodo.5162003},
+      url          = {https://doi.org/10.5281/zenodo.5162003}
     }
 
 To cite a specific version of PyGMT, go to our Zenodo page at
@@ -221,6 +221,7 @@ Documentation for other versions
 * `Development <https://www.pygmt.org/dev>`__ (reflects the *master* branch on
   GitHub)
 * `Latest release <https://www.pygmt.org/latest>`__
+* `v0.4.1 <https://www.pygmt.org/v0.4.1>`__
 * `v0.4.0 <https://www.pygmt.org/v0.4.0>`__
 * `v0.3.1 <https://www.pygmt.org/v0.3.1>`__
 * `v0.3.0 <https://www.pygmt.org/v0.3.0>`__
