+++ doc/changes.md
@@ -39,12 +39,12 @@
 * [Meghan Jones](https://github.com/meghanrjones)
 * [Will Schlitzer](https://github.com/willschlitzer)
 * [Wei Ji Leong](https://github.com/weiji14)
+* [Jiayuan Yao](https://github.com/core-man)
 * [Dongdong Tian](https://github.com/seisman)
 * [Kadatatlu Kishore](https://github.com/kadatatlukishore)
 * [Michael Grund](https://github.com/michaelgrund)
 * [@sean0921](https://github.com/sean0921)
 * [Soham Banerjee](https://github.com/soham4abc)
-* [Jiayuan Yao](https://github.com/core-man)
 * [@yohaimagen](https://github.com/yohaimagen)
 
 ----
