### 2021-08-05 6d62c8f
updating test_xyz2grd.py
