### 2021-08-05 97ece78
add outgrid set test to test_xyz2grd.py
