+++ pygmt/tests/test_xyz2grd.py
@@ -1,11 +1,14 @@
 """
 Tests for xyz2grd.
 """
+import os
+
+import numpy as np
 import pytest
 import xarray as xr
-from pygmt import xyz2grd, grdinfo
+from pygmt import grdinfo, xyz2grd
 from pygmt.datasets import load_sample_bathymetry
-import numpy as np
+from pygmt.helpers import GMTTempFile
 
 
 @pytest.fixture(scope="module", name="ship_data")
@@ -26,6 +29,7 @@ def test_xyz2grd_input_file():
     assert output.gmt.gtype == 0  # Cartesian type
     return output
 
+
 def test_xyz2grd_input_array(ship_data):
     """
     Run xyz2grd by passing in a numpy array.
@@ -36,6 +40,7 @@ def test_xyz2grd_input_array(ship_data):
     assert output.gmt.gtype == 0  # Cartesian type
     return output
 
+
 def test_xyz2grd_input_df(ship_data):
     """
     Run xyz2grd by passing in a data frame.
@@ -45,3 +50,21 @@ def test_xyz2grd_input_df(ship_data):
     assert output.gmt.registration == 0  # Gridline registration
     assert output.gmt.gtype == 0  # Cartesian type
     return output
+
+
+def test_xyz2grd_input_array_file_out(ship_data):
+    """
+    Run xyz2grd by passing in a numpy array and set an outgrid file.
+    """
+    with GMTTempFile(suffix=".nc") as tmpfile:
+        result = xyz2grd(
+            table=np.array(ship_data),
+            spacing=5,
+            region=[245, 255, 20, 30],
+            outgrid=tmpfile.name,
+        )
+        assert result is None  # return value is None
+        assert os.path.exists(path=tmpfile.name)
+        result = grdinfo(tmpfile.name, per_column=True).strip()
+        print(result)
+        assert result == "245 255 20 30 -3651.06079102 -352.379486084 5 5 3 3 0 0"
