### 2021-08-05 3e5f7df
remove kind check in xyz2grd.py
