+++ pygmt/src/xyz2grd.py
@@ -20,14 +20,7 @@ from pygmt.helpers import (
     G="outgrid",
     I="spacing",
     R="region",
-    S="zfile",
-    V="level",
-    Z="single_column",
-    bi="binary",
-    di="nodata",
-    f="flags",
-    h="headers",
-    i="select_column",
+    V="verbose",
 )
 @kwargs_to_strings(R="sequence")
 def xyz2grd(table, **kwargs):
@@ -56,107 +49,7 @@ def xyz2grd(table, **kwargs):
         in.
     {I}
     {R}
-    -A[f|l|m|n|r|s|u|z]
-    By default we will calculate mean values if multiple entries fall on
-    the same node. Use **-A** to change this behavior, except it is
-    ignored if **-Z** is given. Append **f** or **s** to simply keep the
-    first or last data point that was assigned to each node. Append
-    **l** or **u** or **d** to find the lowest (minimum) or upper (maximum)
-    value or the difference between the maximum and miminum value at each
-    node, respectively. Append **m** or **r** or **S** to compute mean or
-    RMS value or standard deviation at each node, respectively. Append **n**
-    to simply count the number of data points that were assigned to each node
-    (this only requires two input columns *x* and *y* as *z* is not consulted).
-    Append **z** to sum multiple values that belong to the same node.
-
-    -Dxname/yname/zname/scale/offset/invalid/title/remark
-    Give values for xname, yname, zname (give the names of those variables and
-    in square bracket their units, e.g., “distance [km]”), scale (to multiply
-    grid values after read [normally 1]), offset (to add to grid after
-    scaling [normally 0]), invalid (a value to represent missing data [NaN]),
-    title (anything you like), and remark (anything you like). To leave some
-    of these values untouched, leave field blank. Empty fields in the end may
-    be skipped. Alternatively, to allow “/” to be part of one of the values,
-    use any non-alphanumeric character (and not the equal sign) as separator
-    by both starting and ending with it.
-
-    -S[zfile]
-    Swap the byte-order of the input only. No grid file is produced.
-    You must also supply the -Z option. The output is written to zfile
-    (or stdout if not supplied).
-    {V}
-
-    -Z[flags]
-    Read a 1-column ASCII [or binary] table. This assumes that all the
-    nodes are present and sorted according to specified ordering
-    convention contained in *flags*. If incoming data represents rows,
-    make *flags* start with **T**\ (op) if first row is y
-    = ymax or **B**\ (ottom) if first row is y = ymin.
-    Then, append **L** or **R** to indicate that first element is at
-    left or right end of row. Likewise for column formats: start with
-    **L** or **R** to position first column, and then append **T** or
-    **B** to position first element in a row. **Note**: These two row/column
-    indicators are only required for grids; for other tables they do not
-    apply. For gridline registered grids: If data are periodic in x but
-    the incoming data do not contain the (redundant) column at x = xmax,
-    append **x**. For data periodic in y without redundant row at y =
-    ymax, append **y**. Append **s**\ *n* to skip the first *n* number
-    of bytes (probably a header). If the byte-order or the words needs
-    to be swapped, append **w**. Select one of several data types (all
-    binary except **a**):
-
-    **A** ASCII representation of one or more floating point values per record
-
-    **a** ASCII representation of a single item per record
-
-    **c** int8_t, signed 1-byte character
-
-    **u** uint8_t, unsigned 1-byte character
-
-    **h** int16_t, signed 2-byte integer
-
-    **H** uint16_t, unsigned 2-byte integer
-
-    **i** int32_t, signed 4-byte integer
-
[…]
