+++ pygmt/src/xyz2grd.py
@@ -3,12 +3,9 @@ xyz2grd - Convert data table to a grid.
 """
 import xarray as xr
 from pygmt.clib import Session
-from pygmt.exceptions import GMTInvalidInput
 from pygmt.helpers import (
     GMTTempFile,
     build_arg_string,
-    data_kind,
-    dummy_context,
     fmt_docstring,
     kwargs_to_strings,
     use_alias,
