### 2021-08-05 dd23548
Combine documentation and compatibility sections in README.md
