### 2021-08-05 b021347
Combine columns
Co-authored-by: Wei Ji <weiji14@users.noreply.github.com>
