### 2021-08-06 f6484b5
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
