### 2021-08-06 2c17aad
Apply suggestions from code review
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
