### 2021-08-06 35e66d1
Update pygmt/tests/test_geopandas.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
