### 2021-08-06 45b5e2e
Apply suggestions from code review
Co-authored-by: Yao Jiayuan <coreman.seism@gmail.com>
