### 2021-08-07 38adbb2
shorten docstring
