### 2021-08-07 b5841d6
Combine documentation and compatibility sections in README.rst (#1415)
Condense the main README file a bit by merging the
PyGMT compatibility table with the list of links to old
versions of the documentation.

Co-authored-by: Wei Ji <weiji14@users.noreply.github.com>
Co-authored-by: Yao Jiayuan <coreman.seism@gmail.com>
