### 2021-08-07 a546150
Update AUTHORS.md
Co-authored-by: Yao Jiayuan <coreman.seism@gmail.com>
