### 2021-08-07 acbf339
Add Yohai Magen to citation
