+++ README.rst
@@ -167,6 +167,7 @@ Developers". Feel free to cite our work in your research using the following Bib
                       Newton, Tyler and
                       Anant, Abhishek and
                       Ziebarth, Malte and
+                      Magen, Yohai and
                       Wessel, Paul},
       title        = {{PyGMT: A Python interface for the Generic Mapping Tools}},
       month        = aug,
