+++ doc/changes.md
@@ -7,7 +7,7 @@
 ### Highlights
 
 * 🎉 **Patch release with multiple gallery examples** 🎉
-* Change the default branch name from "master" to "main" to increase inclusivity ([#1360](https://github.com/GenericMappingTools/pygmt/pull/1360))
+* Change default GitHub branch name from "master" to "main" to increase inclusivity ([#1360](https://github.com/GenericMappingTools/pygmt/pull/1360))
 * Add a "[PyGMT Team](https://www.pygmt.org/latest/team.html)" page ([#1308](https://github.com/GenericMappingTools/pygmt/pull/1308))
 
 ### Enhancements
@@ -21,6 +21,7 @@
 
 ### Documentation
 
+* Combine documentation and compatibility sections in README.rst (([#1415](https://github.com/GenericMappingTools/pygmt/pull/1415))
 * Add a gallery example for grdclip ([#1396](https://github.com/GenericMappingTools/pygmt/pull/1396))
 * Add a gallery example for different colormaps in subplots ([#1394](https://github.com/GenericMappingTools/pygmt/pull/1394))
 * Add a gallery example for the contour method ([#1387](https://github.com/GenericMappingTools/pygmt/pull/1387))
