+++ pygmt/src/__init__.py
@@ -33,8 +33,8 @@ from pygmt.src.plot import plot
 from pygmt.src.plot3d import plot3d
 from pygmt.src.rose import rose
 from pygmt.src.solar import solar
-from pygmt.src.subplot import set_panel, subplot
 from pygmt.src.sphinterpolate import sphinterpolate
+from pygmt.src.subplot import set_panel, subplot
 from pygmt.src.surface import surface
 from pygmt.src.text import text_ as text  # "text" is an argument within "text_"
 from pygmt.src.velo import velo
