### 2021-08-07 31980e8
Changelog entry for v0.4.1 (#1414)
* Add changelog entry for v0.4.1
* Add team page to highlight
* Add change from master to main to highlight
* Add Yohai Magen to citation

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
