+++ pygmt/datasets/__init__.py
@@ -6,6 +6,7 @@ from pygmt.datasets.earth_relief import load_earth_relief
 from pygmt.datasets.samples import (
     load_fractures_compilation,
     load_japan_quakes,
+    load_mars_relief,
     load_ocean_ridge_points,
     load_sample_bathymetry,
     load_usgs_quakes,
