+++ pygmt/datasets/samples.py
@@ -127,7 +127,12 @@ def load_fractures_compilation():
 
 def load_mars_relief():
     """
-    This is the ``@mars370d.txt`` dataset used in the GMT tutorials.
+    Load a table of data for the shape of Mars.
+
+    This is the ``@mars370d.txt`` dataset used in GMT examples, with data and
+    information from Smith, D. E., and M. T. Zuber (1996), The shape of Mars
+    and the topographic signature of the hemispheric dichotomy. Data columns
+    are "longitude," "latitude", and "radius (meters)."
 
     The data are downloaded to a cache directory (usually ``~/.gmt/cache``) the
     first time you invoke this function. Afterwards, it will load the data from
@@ -140,5 +145,5 @@ def load_mars_relief():
         columns.
     """
     fname = which("@mars370d.txt", download="c")
-    data = pd.read_csv(fname, sep=r"\t", header=None, names=["x", "y", "z"])
+    data = pd.read_csv(fname, sep="\t", header=None, names=["lon", "lat", "radius(m)"])
     return data
