### 2021-08-08 43212d2
add test_mars_shape()
