+++ pygmt/tests/test_datasets_samples.py
@@ -4,6 +4,7 @@ Test basic functionality for loading sample datasets.
 from pygmt.datasets import (
     load_fractures_compilation,
     load_japan_quakes,
+    load_mars_shape,
     load_ocean_ridge_points,
     load_sample_bathymetry,
     load_usgs_quakes,
@@ -72,3 +73,17 @@ def test_fractures_compilation():
     assert summary.loc["max", "length"] == 984.652
     assert summary.loc["min", "azimuth"] == 0.0
     assert summary.loc["max", "azimuth"] == 360.0
+
+def test_mars_shape():
+    """
+    Check that the @mars370d.txt dataset loads without errors.
+    """
+    data = load_mars_shape()
+    assert data.shape == (370, 3)
+    summary = data.describe()
+    assert summary.loc["min", "lon"] == 0.008
+    assert summary.loc["max", "lon"] == 359.983
+    assert summary.loc["min", "lat"] == -79.715
+    assert summary.loc["max", "lat"] == 85.887
+    assert summary.loc["min", "radius(m)"] == -6930
+    assert summary.loc["max", "radius(m)"] == 15001
