### 2021-08-08 3fe9fe8
add tests for sphinterpolate
