+++ pygmt/tests/test_sphinterpolate.py
@@ -1,0 +1,33 @@
+"""
+Tests for sphinterpolate.
+"""
+import os
+
+import numpy as np
+import pytest
+from pygmt import grdinfo, sphinterpolate
+from pygmt.exceptions import GMTInvalidInput
+from pygmt.helpers import GMTTempFile
+
+
+def test_sphinterpolate_outgrid():
+    """
+    Test sphinterpolate with a set outgrid.
+    """
+    with GMTTempFile(suffix=".nc") as tmpfile:
+        result = sphinterpolate(
+            table="@mars370d.txt", outgrid=tmpfile.name, increment=1, region="g"
+        )
+        assert result is None  # return value is None
+        assert os.path.exists(path=tmpfile.name)  # check that outgrid exists
+
+
+def test_sphinterpolate_no_outgrid():
+    """
+    Test sphinterpolate with no set outgrid.
+    """
+    temp_grid = sphinterpolate(table="@mars370d.txt", increment=1, region="g")
+    assert temp_grid.dims == ("lat", "lon")
+    assert temp_grid.gmt.gtype == 1  # Geographic grid
+    assert temp_grid.gmt.registration == 0  # Gridline registration
+    result = grdinfo(grid=temp_grid, force_scan="a", per_column="n").strip().split()
