### 2021-08-08 bcb0d2d
add tests in test_sphinterpolate.py
