### 2021-08-08 0318bf5
remove unused imports in test_sphinterpolate.py
