+++ pygmt/tests/test_sphinterpolate.py
@@ -5,9 +5,7 @@ import os
 
 import numpy as np
 import numpy.testing as npt
-import pytest
 from pygmt import grdinfo, sphinterpolate
-from pygmt.exceptions import GMTInvalidInput
 from pygmt.helpers import GMTTempFile
 
 
