### 2021-08-08 23787e8
remove unused imports in test_sphinterpolate.py
