+++ pygmt/src/xyz2grd.py
@@ -31,7 +31,6 @@ def xyz2grd(table, **kwargs):
 
     Full option list at :gmt-docs:`xyz2grd.html`
 
-
     Parameters
     ----------
     table : str or {table-like}
@@ -39,17 +38,20 @@ def xyz2grd(table, **kwargs):
         {table-classes}.
 
     outgrid : str or None
-        The name of the output netCDF file with extension .nc to store the grid
-        in.
+        Optional. The name of the output netCDF file with extension .nc to
+        store the grid in.
     {I}
     {R}
     {V}
+
     Returns
     -------
     ret: xarray.DataArray or None
-    Return type depends on whether the *outgrid* parameter is set:
-    - xarray.DataArray if *outgrid* is not set
-    - None if *outgrid* is set (grid output will be stored in *outgrid*)
+        Return type depends on whether the ``outfile`` parameter is set:
+
+        - :class:`xarray.DataArray`: if ``outfile`` is not set
+        - None if ``outfile`` is set (grid output will be stored in file set by
+          ``outfile``)```
     """
     with GMTTempFile(suffix=".nc") as tmpfile:
         with Session() as lib:
