### 2021-08-08 4727aae
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
