+++ pygmt/tests/test_grdsample.py
@@ -24,13 +24,12 @@ def test_grdsample_file_out(grid):
     grdsample with an outgrid set and the increment is changed.
     """
     with GMTTempFile(suffix=".nc") as tmpfile:
-        result = grdsample(grid=grid, outgrid=tmpfile.name, increment=[1, 2.5])
+        result = grdsample(grid=grid, outgrid=tmpfile.name, increment=[1, 0.5])
         assert result is None  # return value is None
         assert os.path.exists(path=tmpfile.name)  # check that outgrid exists
         result = grdinfo(tmpfile.name, per_column=True).strip().split()
-        print(result)
         assert float(result[6]) == 1  # x-increment
-        assert float(result[7]) == 2.5  # y-increment
+        assert float(result[7]) == 0.5  # y-increment
 
 
 def test_grdsample_no_outgrid(grid):
