### 2021-08-08 10f2474
Update pygmt/src/xyz2grd.py
