### 2021-08-09 097c807
Wrap xyz2grd (#636)
*Wrap the GMT module xyz2grd
*Add tests for xyz2grd

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
Co-authored-by: Wei Ji <weiji.leong@vuw.ac.nz>
Co-authored-by: Will Schlitzer <schlitzer90@gmail.com>
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
