### 2021-08-09 e4f8584
Update pygmt/datasets/samples.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
