### 2021-08-09 85d78d6
Wrap grdsample (#1380)
*Wrap GMT module grdsample
*Add tests for grdsample

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
