### 2021-08-09 275d56c
Add area_thresh to COMMON_OPTIONS
