### 2021-08-09 e6f6514
Update pygmt/helpers/decorators.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
