### 2021-08-09 915ba20
Update pygmt/helpers/decorators.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
