+++ pygmt/helpers/decorators.py
@@ -28,7 +28,7 @@ COMMON_OPTIONS = {
     "A": r"""
         area_thresh : int, float, or str
             *min_area*\ [/*min_level*/*max_level*][**+a**\[**g**\|\ **i**]\
-            [**s**\|\ **S**][**+l**\|\ **r**][**+p**\ *percent*].
+            [**s**\|\ **S**]][**+l**\|\ **r**][**+p**\ *percent*].
             Features with an area smaller than *min_area* in km\ :sup:`2` or of
             hierarchical level that is lower than *min_level* or higher than
             *max_level* will not be plotted.""",
