### 2021-08-10 c5ecf3a
adding a gallery example to gedgradient
