### 2021-08-10 53ef119
Wrap gmtselect
Initial commit for wrapping the gmtselect function for #1427
which selects data table subsets based on multiple spatial
criteria. Original GMT `gmtselect` documentation is at
https://docs.generic-mapping-tools.org/6.2/gmtselect.html.
Aliased non-common optional parameters reverse (I) and
z_subregion (Z).
