### 2021-08-10 bc4825b
Update pygmt/helpers/decorators.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
