+++ pygmt/helpers/decorators.py
@@ -31,7 +31,8 @@ COMMON_OPTIONS = {
             [**s**\|\ **S**]][**+l**\|\ **r**][**+p**\ *percent*].
             Features with an area smaller than *min_area* in km\ :sup:`2` or of
             hierarchical level that is lower than *min_level* or higher than
-            *max_level* will not be plotted.""",
+            *max_level* will not be plotted [Default is 0/0/4 (all
+            features)].""",
     "B": r"""
         frame : bool or str or list
             Set map boundary
