### 2021-08-10 38d49bc
Update pygmt/helpers/decorators.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
