### 2021-08-10 42f346b
NEP29: Set minimum required version to NumPy 1.18+
Following NEP29 policy. Bumps minimum supported
NumPy version to 1.18 in the setup.py, requirements.txt and environment.yml files. Also update installation documentation and set CI tests to run on NumPy 1.18.
