### 2021-08-10 61a9be3
add value-based checks to test_grdproject.py
