+++ pygmt/tests/test_grdproject.py
@@ -4,6 +4,7 @@ Tests for grdproject.
 import os
 
 import pytest
+import numpy.testing as npt
 from pygmt import grdinfo, grdproject
 from pygmt.datasets import load_earth_relief
 from pygmt.exceptions import GMTInvalidInput
@@ -27,6 +28,12 @@ def test_grdproject_file_out(grid):
         assert result is None  # return value is None
         assert os.path.exists(path=tmpfile.name)  # check that outgrid exists
         result = grdinfo(tmpfile.name, per_column=True).strip().split()
+        npt.assert_approx_equal(float(result[0]), 0)  # x min
+        npt.assert_approx_equal(float(result[1]), 10)  # x max
+        npt.assert_approx_equal(float(result[2]), -5.55111512313e-17)  # y min
+        npt.assert_approx_equal(float(result[3]), 9.94585661273)  # y max
+        npt.assert_approx_equal(float(result[4]), -5130.48193359)  # min
+        npt.assert_approx_equal(float(result[5]), -152.585281372)  # max
 
 
 def test_grdproject_no_outgrid(grid):
@@ -38,6 +45,10 @@ def test_grdproject_no_outgrid(grid):
     assert temp_grid.dims == ("y", "x")
     assert temp_grid.gmt.gtype == 0  # Rectangular grid
     assert temp_grid.gmt.registration == 1  # Pixel registration
+    npt.assert_allclose(temp_grid.min(), -5130.482)
+    npt.assert_allclose(temp_grid.max(), -152.58528)
+    npt.assert_allclose(temp_grid.median(), -4578.288)
+    npt.assert_allclose(temp_grid.mean(), -4354.3296)
 
 
 def test_grdproject_fails(grid):
