+++ pygmt/figure.py
@@ -422,8 +422,8 @@ class Figure:
         grdcontour,
         grdimage,
         grdview,
-        hlines,
         histogram,
+        hlines,
         image,
         inset,
         legend,
