+++ pygmt/src/__init__.py
@@ -22,8 +22,8 @@ from pygmt.src.grdlandmask import grdlandmask
 from pygmt.src.grdsample import grdsample
 from pygmt.src.grdtrack import grdtrack
 from pygmt.src.grdview import grdview
-from pygmt.src.hlines import hlines
 from pygmt.src.histogram import histogram
+from pygmt.src.hlines import hlines
 from pygmt.src.image import image
 from pygmt.src.info import info
 from pygmt.src.inset import inset
