### 2021-08-10 f2c84a7
adjust priliminary tests
