### 2021-08-10 b0d2901
add ternary.py
