+++ pygmt/figure.py
@@ -434,6 +434,7 @@ class Figure:
         set_panel,
         solar,
         subplot,
+        ternary,
         text,
         velo,
         wiggle,
