+++ pygmt/src/sphdistance.py
@@ -8,8 +8,6 @@ from pygmt.exceptions import GMTInvalidInput
 from pygmt.helpers import (
     GMTTempFile,
     build_arg_string,
-    data_kind,
-    dummy_context,
     fmt_docstring,
     kwargs_to_strings,
     use_alias,
@@ -54,15 +52,9 @@ def sphdistance(table, **kwargs):
     """
     if "I" not in kwargs.keys() or "R" not in kwargs.keys():
         raise GMTInvalidInput("Both 'region' and 'increment' must be specified.")
-    kind = data_kind(table)
     with GMTTempFile(suffix=".nc") as tmpfile:
         with Session() as lib:
-            if kind == "file":
-                file_context = dummy_context(table)
-            elif kind == "matrix":
-                file_context = lib.virtualfile_from_matrix(matrix=table)
-            else:
-                raise GMTInvalidInput("Unrecognized data type: {}".format(type(table)))
+            file_context = lib.virtualfile_from_data(check_kind="vector", data=table)
             with file_context as infile:
                 if "G" not in kwargs.keys():  # if outgrid is unset, output to tempfile
                     kwargs.update({"G": tmpfile.name})
