+++ pygmt/tests/test_sph2grd.py
@@ -1,0 +1,37 @@
+"""
+Tests for sph2grd.
+"""
+import os
+
+import numpy.testing as npt
+from pygmt import grdinfo, sph2grd
+from pygmt.helpers import GMTTempFile
+
+
+def test_sph2grd_outgrid():
+    """
+    Test sph2grd with a set outgrid.
+    """
+    with GMTTempFile(suffix=".nc") as tmpfile:
+        result = sph2grd(
+            table="@EGM96_to_36.txt", outgrid=tmpfile.name, increment=1, region="g"
+        )
+        assert result is None  # return value is None
+        assert os.path.exists(path=tmpfile.name)  # check that outgrid exists
+        result = grdinfo(tmpfile.name, per_column=True).strip().split()
+        assert int(result[0]) == 0  # x minimum
+        assert int(result[1]) == 360  # x maximum
+        assert int(result[2]) == -90  # y minimum
+        assert int(result[3]) == 90  # y maximum
+        npt.assert_approx_equal(float(result[4]), -0.00043260390521)  # v minimum
+        npt.assert_approx_equal(float(result[5]), 0.000219614637899)  # v maximum
+
+
+def test_sph2grd_no_outgrid():
+    """
+    Test sph2grd with no set outgrid.
+    """
+    temp_grid = sph2grd(table="@EGM96_to_36.txt", increment=1, region="g")
+    assert temp_grid.dims == ("y", "x")
+    assert temp_grid.gmt.gtype == 0  # Cartesian grid
+    assert temp_grid.gmt.registration == 0  # Gridline registration
