### 2021-08-10 2e033c5
add resolution parameter docstring
