+++ pygmt/src/grdlandmask.py
@@ -49,8 +49,18 @@ def grdlandmask(**kwargs):
         in.
     {I}
     {R}
-    area_thresh: todo
-    resolution : todo
+    {A}
+    resolution : str
+        res[**+f**]. Selects the resolution of the data set to use
+        ((**f**)ull, (**h**)igh, (**i**)ntermediate, (**l**)ow, or
+        (**c**)rude). The resolution drops off by ~80% between data sets.
+        [Default is **l**]. Append **+f** to automatically select a lower
+        resolution should the one requested not be available
+        [abort if not found]. Alternatively, choose (**a**)uto to automatically
+        select the best resolution given the chosen region. Note that because
+        the coastlines differ in details a node in a mask file using one
+        resolution is not guaranteed to remain inside [or outside] when a
+        different resolution is selected.
     bordervalues : todo
     mask_geog : todo
     {V}
