### 2021-08-10 89789dd
add bordervalues parameter docstring
