+++ pygmt/src/grdlandmask.py
@@ -22,7 +22,7 @@ from pygmt.helpers import (
     A="area_thresh",
     D="resolution",
     E="bordervalues",
-    N="mask_geog",
+    N="maskvalues",
     V="verbose",
     r="registration",
 )
@@ -51,7 +51,7 @@ def grdlandmask(**kwargs):
     {R}
     {A}
     resolution : str
-        res[**+f**]. Selects the resolution of the data set to use
+        *res*[**+f**]. Selects the resolution of the data set to use
         ((**f**)ull, (**h**)igh, (**i**)ntermediate, (**l**)ow, or
         (**c**)rude). The resolution drops off by ~80% between data sets.
         [Default is **l**]. Append **+f** to automatically select a lower
@@ -62,7 +62,7 @@ def grdlandmask(**kwargs):
         resolution is not guaranteed to remain inside [or outside] when a
         different resolution is selected.
     bordervalues : str or list
-        [values]. Nodes that fall exactly on a polygon boundary should be
+        [*values*]. Nodes that fall exactly on a polygon boundary should be
         considered to be outside the polygon [Default considers them to be
         inside]. Alternatively, append either the four values *cborder*/
         *lborder*/*iborder*/*pborder* or just the single value *bordervalue*
@@ -74,7 +74,19 @@ def grdlandmask(**kwargs):
         cells traversed by a lake outline, *iborder* for islands-in-lakes
         outlines, and *pborder* for ponds-in-islands-in-lakes outlines
         [Default is no line tracing].
-    mask_geog : todo
+    maskvalues : str or list
+        *values*. Sets the values that will be assigned to nodes. Values can
+        be any number, including the textstring NaN. Also select
+        ``bordervalues``` to let nodes exactly on feature boundaries be
+        considered outside [Default is inside]. Specify this information
+        using 1 of 2 formats:
+
+        -``maskvalues``*wet*/*dry*.
+
+        -``maskvalues``*ocean*/*land*/*lake*/*island*/*pond*.
+
+        [Default is 0/1/0/1/0 (i.e., 0/1)].
+
     {V}
     {r}
 
