### 2021-08-10 d1d359b
List NumPy >= 1.18 in compatibility table on main README.rst
Also link to v0.5.0 milestone and development documentation page.
