### 2021-08-11 3f4090e
Refactor triangulate to use virtualfile_from_data
Modernizing the code implementation to keep up to date
with PyGMT v0.4.1. Also refreshing the unit tests a bit.
