+++ pygmt/tests/test_triangulate.py
@@ -6,122 +6,94 @@ import os
 import pandas as pd
 import pytest
 import xarray as xr
-from pygmt import triangulate, which
+from pygmt import triangulate
 from pygmt.datasets import load_sample_bathymetry
 from pygmt.exceptions import GMTInvalidInput
-from pygmt.helpers import data_kind
-
-TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
-TEMP_GRID = os.path.join(TEST_DATA_DIR, "tmp_grid.nc")
+from pygmt.helpers import GMTTempFile, data_kind
 
 
-@pytest.fixture(scope="module", name="ship_data")
-def fixture_ship_data():
+@pytest.fixture(scope="module", name="dataframe")
+def fixture_dataframe():
     """
-    Load the grid data from the sample bathymetry file.
+    Load the table data from the sample bathymetry dataset.
     """
-    ship_data = load_sample_bathymetry()
-    return ship_data
+    return load_sample_bathymetry()
 
 
 def test_triangulate_input_file():
     """
     Run triangulate by passing in a filename.
     """
-    fname = which("@tut_ship.xyz", download="c")
-    output = triangulate(data=fname)
+    output = triangulate(table="@tut_ship.xyz")
     assert isinstance(output, pd.DataFrame)
     assert output.shape == (161935, 3)
-    return output
 
 
-def test_triangulate_input_data_array(ship_data):
+def test_triangulate_input_data_array(dataframe):
     """
     Run triangulate by passing in a numpy array into data.
     """
-    data = ship_data.to_numpy()
-    output = triangulate(data=data)
+    data = dataframe.to_numpy()
+    output = triangulate(table=data)
     assert isinstance(output, pd.DataFrame)
     assert output.shape == (161935, 3)
-    return output
 
 
-def test_triangulate_input_xyz(ship_data):
+def test_triangulate_input_xyz(dataframe):
     """
     Run triangulate by passing in x, y, z numpy.ndarrays individually.
     """
     output = triangulate(
-        x=ship_data.longitude,
-        y=ship_data.latitude,
-        z=ship_data.bathymetry,
+        x=dataframe.longitude,
+        y=dataframe.latitude,
+        z=dataframe.bathymetry,
     )
     assert isinstance(output, pd.DataFrame)
     assert output.shape == (161935, 3)
-    return output
 
 
-def test_triangulate_input_xy_no_z(ship_data):
+def test_triangulate_input_xy_no_z(dataframe):
     """
     Run triangulate by passing in x and y, but no z.
     """
-    with pytest.raises(GMTInvalidInput):
-        triangulate(x=ship_data.longitude, y=ship_data.latitude)
+    output = triangulate(x=dataframe.longitude, y=dataframe.latitude)
+    assert isinstance(output, pd.DataFrame)
+    assert output.shape == (161935, 3)
 
 
-def test_triangulate_wrong_kind_of_input(ship_data):
+def test_triangulate_wrong_kind_of_input(dataframe):
     """
     Run triangulate using grid input that is not file/matrix/vectors.
     """
-    data = ship_data.bathymetry.to_xarray()  # convert pandas.Series to xarray.DataArray
+    data = dataframe.bathymetry.to_xarray()  # convert pandas.Series to xarray.DataArray
     assert data_kind(data) == "grid"
     with pytest.raises(GMTInvalidInput):
-        triangulate(data=data)
+        triangulate(table=data)
 
 
-def test_triangulate_with_outgrid_true(ship_data):
+def test_triangulate_with_outgrid_true(dataframe):
     """
     Run triangulate with outgrid=True and see it load into an xarray.DataArray.
     """
-    data = ship_data.to_numpy()
[…]
