+++ examples/gallery/images/grdgradient.py
@@ -1,9 +1,9 @@
 """
 Calculating the gradient of a grid file
 --------------------
-The :meth:`pygmt.grdclip` calculate the gradient of a grid file.
+The :meth:`pygmt.grgradient` calculate the gradient of a grid file.
 In the example shown below we will see how to calculate an hillshade map based on
-a Data Elevation Model(DEM). :meth:`pygmt.grdclip` get as input :class:`xarray.DataArray` object
+a Data Elevation Model(DEM). :meth:`pygmt.grdgradient` get as input :class:`xarray.DataArray` object
 or a path string to a grid file. We will use the ``radiance`` parameter in order to set the illumination
 source direction and altitude
 """
