### 2021-08-11 69a31ec
Wrap grdmix
Initial commit for wrapping the grdmix function for #578
which does blending and transforming of grids and images.
Original GMT `grdmix` documentation is at
https://docs.generic-mapping-tools.org/6.2/grdmix.html.
Aliased non-common optional parameters construct (C),
deconstruct (D) and normalize (N). Added one unit test
to deconstruct the earth_day_01d.tif image into 3 bands.
