### 2021-08-11 72f7e89
adding default behavior for ploting gmt file at plot.py
