+++ pygmt/src/plot.py
@@ -207,6 +207,11 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
     extra_arrays = []
     if "S" in kwargs and kwargs["S"][0] in "vV" and direction is not None:
         extra_arrays.extend(direction)
+    if "S" not in kwargs and kind == "file":
+        with open(data, "r") as f:
+            line = f.readline()
+        if "@GMULTIPOINT" in line or "@GPOINT" in line:
+            kwargs["S"] = "s0.2c"
     if "G" in kwargs and not isinstance(kwargs["G"], str):
         if kind != "vectors":
             raise GMTInvalidInput(
