### 2021-08-11 4a1eac6
adjust ordering in index.rst
