+++ doc/api/index.rst
@@ -92,8 +92,8 @@ Operations on grids:
     grdcut
     grdfill
     grdfilter
-    grdlandmask
     grdgradient
+    grdlandmask
     grdsample
     grdtrack
     xyz2grd
