### 2021-08-11 0b642a9
adding test for non defualt style
