+++ pygmt/tests/test_geopandas.py
@@ -97,3 +97,36 @@ def test_geopandas_plot3d_default_cube():
         zscale=1.5,
     )
     return fig
+
+@pytest.mark.mpl_image_compare
+def test_geopandas_plot_non_default_circle():
+    """
+    Check the default behavior of plotting geopandas DataFrame with Point
+    geometry in 2d.
+    """
+    point = shapely.geometry.Point(1, 2)
+    gdf = gpd.GeoDataFrame(geometry=[point])
+    fig = Figure()
+    fig.plot(data=gdf, region=[0, 2, 1, 3], projection="X2c", frame=True, style='c0.2c')
+    return fig
+
+
+@pytest.mark.mpl_image_compare
+def test_geopandas_plot3d_non_default_circle():
+    """
+    Check the default behavior of plotting geopandas DataFrame with MultiPoint
+    geometry in 3d.
+    """
+    multipoint = shapely.geometry.MultiPoint([(0.5, 0.5, 0.5), (1.5, 1.5, 1.5)])
+    gdf = gpd.GeoDataFrame(geometry=[multipoint])
+    fig = Figure()
+    fig.plot3d(
+        data=gdf,
+        perspective=[315, 25],
+        region=[0, 2, 0, 2, 0, 2],
+        projection="X2c",
+        frame=["WsNeZ1", "xag", "yag", "zag"],
+        zscale=1.5,
+        style='c0.2c'
+    )
+    return fig
