+++ pygmt/src/hlines.py
@@ -74,7 +74,8 @@ def hlines(self, y=None, xmin=None, xmax=None, **kwargs):
 
         if xmin is None and xmax is None:
             with Session() as lib:
-                # get limits from current map boundings if not given via xmin, xmax
+                # get limits from current map boundings if not given
+                # via xmin, xmax
                 x = np.array([[lib.extract_region()[0]], [lib.extract_region()[1]]])
                 x = np.repeat(x, list_length, axis=1)
         elif xmin is None or xmax is None:
@@ -83,7 +84,8 @@ def hlines(self, y=None, xmin=None, xmax=None, **kwargs):
             )
 
         else:
-            # if only a single xmin and xmax without [], repeat to fit size of y
+            # if only a single xmin and xmax without [], repeat to fit
+            # size of y
             if isinstance(xmin, (int, float)):
                 x = np.array([[xmin], [xmax]])
                 x = np.repeat(x, list_length, axis=1)
