### 2021-08-11 37d27c5
Create process_output_grid function in pygmt/io.py
