### 2021-08-11 2214237
Use process_output_grid() to return dataarray or None
