+++ pygmt/src/xyz2grd.py
@@ -1,7 +1,6 @@
 """
 xyz2grd - Convert data table to a grid.
 """
-import xarray as xr
 from pygmt.clib import Session
 from pygmt.helpers import (
     GMTTempFile,
@@ -10,6 +9,7 @@ from pygmt.helpers import (
     kwargs_to_strings,
     use_alias,
 )
+from pygmt.io import process_output_grid
 
 
 @fmt_docstring
@@ -64,11 +64,4 @@ def xyz2grd(table, **kwargs):
                 arg_str = " ".join([infile, arg_str])
                 lib.call_module("xyz2grd", arg_str)
 
-        if outgrid == tmpfile.name:  # if user did not set outgrid, return DataArray
-            with xr.open_dataarray(outgrid) as dataarray:
-                result = dataarray.load()
-                _ = result.gmt  # load GMTDataArray accessor information
-        else:
-            result = None  # if user sets an outgrid, return None
-
-        return result
+        return process_output_grid(outgrid, tmpfile.name)
