### 2021-08-11 ab8954e
Update pygmt/io.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
