+++ pygmt/io.py
@@ -1,5 +1,5 @@
 """
-PyGMT i/o utilities.
+PyGMT input/output (I/O) utilities.
 """
 import xarray as xr
 
