### 2021-08-12 5d778f1
adding a coment explaining the if statment
