### 2021-08-13 6456fac
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
