### 2021-08-14 96b683c
Update pygmt/src/plot.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
