### 2021-08-14 e57ec91
Update pygmt/src/plot3d.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
