### 2021-08-17 b7827e0
Update pygmt/tests/test_geopandas.py
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
