### 2021-08-17 cd822ca
Plot square or cube by default for geopandas Point/MultiPoint types (#1405)
Set default behavior for plotting geopandas dataframe with
Point/Multipoint type geometry as points (-Sp0.2c) for `plot`
and cubes (-Su0.2c) for `plot3d`, instead of GMT's default
line style plot.

* testing default behavior of plot and plot3d with geopandas Point and MultiPoint DataFrames
* adding test for non default style
* adding a comment explaining the if statment

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
Co-a
[…]
