### 2021-08-17 8a20e99
checks that the file is not on server
