### 2021-08-17 342efa7
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
