### 2021-08-18 b31162e
Update pygmt/src/grdlandmask.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
