### 2021-08-18 2d3e197
changing the if statmrnt to try except style
