### 2021-08-18 c9e525f
add the same if statment to plot3d.py
