+++ pygmt/src/plot3d.py
@@ -177,6 +177,18 @@ def plot3d(
     extra_arrays = []
     if "S" in kwargs and kwargs["S"][0] in "vV" and direction is not None:
         extra_arrays.extend(direction)
+    if (
+        "S" not in kwargs and kind == "file"
+    ):  # chacking that the data is a file path to set defualt style
+        try:
+            with open(pygmt.which(data), "r") as file:
+                line = file.readline()
+            if (
+                "@GMULTIPOINT" in line or "@GPOINT" in line
+            ):  # if the file is gmt style and geometry is set to Point
+                kwargs["S"] = "u0.2c"
+        except FileNotFoundError:
+            pas
     if "G" in kwargs and not isinstance(kwargs["G"], str):
         if kind != "vectors":
             raise GMTInvalidInput(
