### 2021-08-18 1ca9b0d
Add missing aliases to pygmt.Figure.image.py
