### 2021-08-18 5695605
Add missing aliases to pygmt.Figure.text
