### 2021-08-18 8be5606
Add missing aliases to pygmt.Figure.histogram
