### 2021-08-19 b797861
Update pygmt/src/plot.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
