### 2021-08-19 c0a355c
Update pygmt/src/plot.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
