### 2021-08-19 6e29ba6
Update pygmt/src/plot3d.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
