### 2021-08-19 9d62ce9
Update pygmt/src/plot3d.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
