### 2021-08-19 befae12
Update pygmt/src/contour.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
