+++ pygmt/src/contour.py
@@ -36,8 +36,8 @@ from pygmt.helpers import (
     d="nodata",
     e="find",
     f="coltypes",
-    i="incols",
     h="header",
+    i="incols",
     l="label",
     p="perspective",
     t="transparency",
