+++ pygmt/src/contour.py
@@ -112,18 +112,18 @@ def contour(self, x=None, y=None, z=None, data=None, **kwargs):
         to be of the format [*annotcontlabel*][/*contlabel*]. If either
         label contains a slash (/) character then use ``|`` as the
         separator for the two labels instead.
+    {U}
     {V}
     {XY}
-    {c}
-    {i}
-    {p}
-    {t}
     {b}
+    {c}
     {d}
     {e}
     {f}
     {h}
-    {U}
+    {i}
+    {p}
+    {t}
     """
     kwargs = self._preprocess(**kwargs)  # pylint: disable=protected-access
 
