### 2021-08-19 516f960
add tests to test_plot.py
