+++ pygmt/src/plot3d.py
@@ -12,6 +12,7 @@ from pygmt.helpers import (
     kwargs_to_strings,
     use_alias,
 )
+from pygmt.src.which import which
 
 
 @fmt_docstring
@@ -178,9 +179,9 @@ def plot3d(
     if "S" in kwargs and kwargs["S"][0] in "vV" and direction is not None:
         extra_arrays.extend(direction)
     elif (
-            "S" not in kwargs
-            and kind == "geojson"
-            and data.geom_type.isin(["Point", "MultiPoint"]).all()
+        "S" not in kwargs
+        and kind == "geojson"
+        and data.geom_type.isin(["Point", "MultiPoint"]).all()
     ):  # checking if the geometry of a geoDataFrame is Point or MultiPoint
         kwargs["S"] = "u0.2c"
     elif (
