### 2021-08-19 029a852
adding tests to test_plot3d.py
