### 2021-08-19 8897946
Expand table-like input options for pygmt.surface
