### 2021-08-19 852e0cf
Update pygmt/datasets/earth_relief.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
