### 2021-08-19 e55805b
Initial load earth night function
