### 2021-08-19 c8cfb28
Update pygmt/src/surface.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
