### 2021-08-19 79e3edf
Load earth_night_01d
