+++ pygmt/src/surface.py
@@ -86,7 +86,7 @@ def surface(x=None, y=None, z=None, data=None, **kwargs):
         with Session() as lib:
             # Choose how data will be passed into the module
             file_context = lib.virtualfile_from_data(
-                    check_kind="vector", data=data, x=x, y=y, z=z
+                check_kind="vector", data=data, x=x, y=y, z=z
             )
             with file_context as infile:
                 if "G" not in kwargs.keys():  # if outfile is unset, output to tmpfile
