### 2021-08-19 ec4c81d
[format-command] fixes
