### 2021-08-19 d0d81cf
Switch outfile to outgrid in test file
