+++ pygmt/tests/test_surface.py
@@ -82,17 +82,17 @@ def test_surface_wrong_kind_of_input(ship_data):
         surface(data=data, spacing="5m", region=[245, 255, 20, 30])
 
 
-def test_surface_with_outfile_param(ship_data):
+def test_surface_with_outgrid_param(ship_data):
     """
     Run surface with the -Goutputfile.nc parameter.
     """
     data = ship_data.values  # convert pandas.DataFrame to numpy.ndarray
     try:
         output = surface(
-            data=data, spacing="5m", region=[245, 255, 20, 30], outfile=TEMP_GRID
+            data=data, spacing="5m", region=[245, 255, 20, 30], outgrid=TEMP_GRID
         )
-        assert output is None  # check that output is None since outfile is set
-        assert os.path.exists(path=TEMP_GRID)  # check that outfile exists at path
+        assert output is None  # check that output is None since outgrid is set
+        assert os.path.exists(path=TEMP_GRID)  # check that outgrid exists at path
         with xr.open_dataarray(TEMP_GRID) as grid:
             assert isinstance(grid, xr.DataArray)  # ensure netcdf grid loads ok
     finally:
@@ -103,13 +103,13 @@ def test_surface_with_outfile_param(ship_data):
 def test_surface_short_aliases(ship_data):
     """
     Run surface using short aliases -I for spacing, -R for region, -G for
-    outfile.
+    outgrid.
     """
     data = ship_data.values  # convert pandas.DataFrame to numpy.ndarray
     try:
         output = surface(data=data, I="5m", R=[245, 255, 20, 30], G=TEMP_GRID)
-        assert output is None  # check that output is None since outfile is set
-        assert os.path.exists(path=TEMP_GRID)  # check that outfile exists at path
+        assert output is None  # check that output is None since outgrid is set
+        assert os.path.exists(path=TEMP_GRID)  # check that outgrid exists at path
         with xr.open_dataarray(TEMP_GRID) as grid:
             assert isinstance(grid, xr.DataArray)  # ensure netcdf grid loads ok
     finally:
