### 2021-08-20 8becd43
Add missing alias to pygmt.Figure.legend (#1453)
Add in missing alias info for U - timestamp to pygmt.Figure.legend
