### 2021-08-22 8089ef9
update load_hotspots() for new format of source file
