### 2021-08-22 a028634
Fix lint warnings for pylint 2.10 (#1459)
