+++ pygmt/tests/test_legend.py
@@ -93,7 +93,7 @@ T so we may have to adjust the box height to get the right size box.
 """
 
     with GMTTempFile() as specfile:
-        with open(specfile.name, "w") as file:
+        with open(specfile.name, "w", encoding="utf8") as file:
             file.write(specfile_contents)
         fig = Figure()
         fig.basemap(projection="x6i", region=[0, 1, 0, 1], frame=True)
