### 2021-08-22 9dc3fb7
Expand tests to support Windows and POSIX Path classes
