+++ pygmt/tests/test_info.py
@@ -44,6 +44,34 @@ def test_info_path():
     assert output == expected_output
 
 
+def test_info_pure_windows_path():
+    """
+    Make sure info works on a pathlib.Path input.
+    """
+    output = info(table=pathlib.PureWindowsPath(POINTS_DATA))
+    expected_output = (
+        f"{POINTS_DATA}: N = 20 "
+        "<11.5309/61.7074> "
+        "<-2.9289/7.8648> "
+        "<0.1412/0.9338>\n"
+    )
+    assert output == expected_output
+
+
+def test_info_pure_posix_path():
+    """
+    Make sure info works on a pathlib.Path input.
+    """
+    output = info(table=pathlib.PurePosixPath(POINTS_DATA))
+    expected_output = (
+        f"{POINTS_DATA}: N = 20 "
+        "<11.5309/61.7074> "
+        "<-2.9289/7.8648> "
+        "<0.1412/0.9338>\n"
+    )
+    assert output == expected_output
+
+
 def test_info_2d_list():
     """
     Make sure info works on a 2d list.
