### 2021-08-23 0b28914
change the name of plot.py tests
