### 2021-08-23 429b34d
changing the names of the plot3d tests
