### 2021-08-23 e2005e5
Update pygmt/src/text.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
