### 2021-08-23 eb9d7f3
Update pygmt/src/text.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
