### 2021-08-23 2cfb303
Add missing aliases to pygmt.Figure.text (#1448)
Adds common options timestamp (U), aspatial (a), find (e),
header (h), incols (i) and wrap (w) to the pygmt.Figure.text method.

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
