### 2021-08-24 dce1d0c
Update pygmt/src/plot.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
