+++ pygmt/src/plot.py
@@ -218,7 +218,7 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
         "S" not in kwargs and kind == "file"
     ):  # checking that the data is a file path to set default style
         try:
-            with open(which(data), "r") as file:
+            with open(which(data), mode="r", encoding="utf8") as file:
                 line = file.readline()
             if (
                 "@GMULTIPOINT" in line or "@GPOINT" in line
