### 2021-08-24 393cd5c
Update pygmt/src/plot3d.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
