+++ pygmt/src/plot3d.py
@@ -188,7 +188,7 @@ def plot3d(
         "S" not in kwargs and kind == "file"
     ):  # checking that the data is a file path to set default style
         try:
-            with open(which(data), "r") as file:
+            with open(which(data), mode="r", encoding="utf8") as file:
                 line = file.readline()
             if (
                 "@GMULTIPOINT" in line or "@GPOINT" in line
