+++ pygmt/tests/test_plot3d.py
@@ -533,7 +533,7 @@ def test_plot3d_ogrgmt_file_multipoint_default_style():
 >
 1 1 2
 1.5 1.5 1"""
-        with open(tmpfile.name, "w") as file:
+        with open(tmpfile.name, "w", encoding="utf8") as file:
             file.write(gmt_file)
         fig = Figure()
         fig.plot3d(
@@ -559,7 +559,7 @@ def test_plot3d_ogrgmt_file_multipoint_non_default_style():
 >
 1 1 2
 1.5 1.5 1"""
-        with open(tmpfile.name, "w") as file:
+        with open(tmpfile.name, "w", encoding="utf8") as file:
             file.write(gmt_file)
         fig = Figure()
         fig.plot3d(
