### 2021-08-24 022b29e
ignors to many varibels at plot.py and plot3d.py
