### 2021-08-24 6999a52
Update pygmt/tests/test_plot.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
