### 2021-08-24 8638cdd
Update pygmt/tests/test_plot3d.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
