### 2021-08-24 9330943
Apply suggestions from code review
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
