### 2021-08-24 c146942
add load_hotspots to docs
