+++ doc/api/index.rst
@@ -159,6 +159,7 @@ and store them in the GMT cache folder.
     datasets.load_sample_bathymetry
     datasets.load_usgs_quakes
     datasets.load_fractures_compilation
+    datasets.load_hotspots
 
 .. automodule:: pygmt.exceptions
 
