### 2021-08-25 74bd867
Update pygmt/tests/test_plot.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
