### 2021-08-25 ebe4549
Update pygmt/tests/test_plot3d.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
