### 2021-08-25 a1e8cf5
Update pygmt/tests/test_plot.py
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
