+++ pygmt/tests/test_plot.py
@@ -501,7 +501,7 @@ def test_plot_deprecate_columns_to_incols(region):
 @pytest.mark.mpl_image_compare
 def test_plot_ogrgmt_file_multipoint_default_style():
     """
-    Make sure that OGR/GMT file with MultiPoint geometry are plotted as squares
+    Make sure that OGR/GMT files with MultiPoint geometry are plotted as squares
     and not as line (default GMT style).
     """
     with GMTTempFile(suffix=".gmt") as tmpfile:
