### 2021-08-25 707dc96
Update pygmt/tests/test_plot3d.py
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
