### 2021-08-25 0881d0f
Plot square or cube by default for OGR/GMT files with Point/MultiPoint types (#1438)
Set default behavior for plotting OGR/GMT files with
Point/Multipoint type geometry as points (-Sp0.2c) for `plot`
and cubes (-Su0.2c) for `plot3d`, instead of GMT's default
line style plot.

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
