+++ pygmt/src/plot.py
@@ -12,6 +12,7 @@ from pygmt.helpers import (
     kwargs_to_strings,
     use_alias,
 )
+from pygmt.src.which import which
 
 
 @fmt_docstring
@@ -200,6 +201,7 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
         *transparency* can also be a 1d array to set varying transparency
         for symbols, but this option is only valid if using x/y.
     """
+    # pylint: disable=too-many-locals
     kwargs = self._preprocess(**kwargs)  # pylint: disable=protected-access
 
     kind = data_kind(data, x, y)
@@ -213,6 +215,18 @@ def plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs):
         and data.geom_type.isin(["Point", "MultiPoint"]).all()
     ):  # checking if the geometry of a geoDataFrame is Point or MultiPoint
         kwargs["S"] = "s0.2c"
+    elif (
+        "S" not in kwargs and kind == "file"
+    ):  # checking that the data is a file path to set default style
+        try:
+            with open(which(data), mode="r", encoding="utf8") as file:
+                line = file.readline()
+            if (
+                "@GMULTIPOINT" in line or "@GPOINT" in line
+            ):  # if the file is gmt style and geometry is set to Point
+                kwargs["S"] = "s0.2c"
+        except FileNotFoundError:
+            pass
     if "G" in kwargs and not isinstance(kwargs["G"], str):
         if kind != "vectors":
             raise GMTInvalidInput(
