### 2021-08-25 29c8be6
Apply suggestions from code review
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
