### 2021-08-25 ba39ef7
update main docstring for grdproject
