+++ pygmt/src/grdvolume.py
@@ -1,7 +1,6 @@
 """
 grdvolume - Calculate grid volume and area constrained by a contour.
 """
-import numpy as np
 import pandas as pd
 from pygmt.clib import Session
 from pygmt.exceptions import GMTInvalidInput
