### 2021-08-25 1639d74
Update pygmt/datasets/samples.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
