### 2021-08-26 8a6db70
Add function to import hotspot dataset (#1386)
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
