### 2021-08-27 a56a281
Add gallery example for grdlandmask
According to #1421 here's now the gallery example showcasing the use of ``grdlandmask``.
