### 2021-08-28 90476cd
Update pygmt/tests/test_grdproject.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
