### 2021-08-28 ffbcdfe
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
