+++ examples/gallery/images/grdlandmask.py
@@ -1,7 +1,7 @@
 """
 Create 'wet-dry' mask grid
---------------------
-The :meth:`pygmt.grdlandmash` method allows to set 
+--------------------------
+The :meth:`pygmt.grdlandmask` method allows setting 
 all nodes on land or water to a specified value using
 the ``maskvalues`` parameter. 
 """
@@ -14,7 +14,7 @@ fig = pygmt.Figure()
 region = [-65, -40, -40, -20]
 
 # Assign a value of 0 for all water masses and a value of 1 for all land masses
-grid = pygmt.grdlandmask(region=region, spacing="5m", maskvalues="0/1", resolution="l")
+grid = pygmt.grdlandmask(region=region, spacing="5m", maskvalues=[0, 1], resolution="l")
 
 # Plot clipped grid
 fig.basemap(region=region, projection="M12c", frame=True)
