### 2021-08-30 d37dac6
Add the 'incols' alias to info
