### 2021-08-31 66dc1ce
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
