### 2021-08-31 6bfd0a3
Update modules to use load_dataarray
