+++ pygmt/src/grdlandmask.py
@@ -11,7 +11,7 @@ from pygmt.helpers import (
     kwargs_to_strings,
     use_alias,
 )
-from pygmt.io import process_output_grid
+from pygmt.io import load_dataarray
 
 
 @fmt_docstring
@@ -68,4 +68,4 @@ def grdlandmask(**kwargs):
             arg_str = build_arg_string(kwargs)
             lib.call_module("grdlandmask", arg_str)
 
-        return process_output_grid(outgrid, tmpfile.name)
+        return load_dataarray(outgrid) if outgrid == tmpfile.name else None
