+++ examples/gallery/images/grdlandmask.py
@@ -14,7 +14,7 @@ fig = pygmt.Figure()
 region = [-65, -40, -40, -20]
 
 # Assign a value of 0 for all water masses and a value of 1 for all land masses.
-# Use shoreline data with (l)ow resolution and set the grid spacing to 
+# Use shoreline data with (l)ow resolution and set the grid spacing to
 # 5 arc minute in x and y direction.
 grid = pygmt.grdlandmask(region=region, spacing="5m", maskvalues=[0, 1], resolution="l")
 
