### 2021-08-31 3da72c1
Update docstring
