+++ pygmt/src/grd2xyz.py
@@ -76,7 +76,9 @@ def grd2xyz(grid, output_type="pandas", outfile=None, **kwargs):
         output_type = "file"
     elif output_type == "file" and outfile is None:
         raise GMTInvalidInput("""Must specify outfile for ASCII output.""")
+    # Set the default column names for the pandas dataframe header
     dataframe_header = ["x", "y", "z"]
+    # Set the column names to match an input DataArray as the grid
     if type(grid) == xr.DataArray and output_type == "pandas":
         if len(grid.dims) == 3:
             dataframe_header = grid.dims
