+++ examples/gallery/lines/linefronts.py
@@ -49,25 +49,30 @@ for frontstyle in [
     "f1c/0.25c+c",
     # line with triangle front style
     "f1c/0.3c+t",
-    # line with left-lateral ("+l") slip ("+s") front style, angle is set to 45 and
-    # offset to 2.25 cm
+    # line with left-lateral ("+l") slip ("+s") front style, angle is set to 45
+    # and offset to 2.25 cm
     "f5c/1c+l+s45+o2.25c",
-    # line with "faults" front style, symbols are plotted on the left side of the front
+    # line with "faults" front style, symbols are plotted on the left side of
+    # the front
     "f1c/0.4c+l",
-    # line with box front style, symbols are plotted on the left side of the front
+    # line with box front style, symbols are plotted on the left side of the
+    # front
     "f1c/0.3c+l+b",
-    # line with circle front style, symbols are plotted on the right side of the front
+    # line with circle front style, symbols are plotted on the right side of
+    # the front
     "f1c/0.4c+r+c",
-    # line with triangle front style, symbols are plotted on the left side of the front
+    # line with triangle front style, symbols are plotted on the left side of
+    # the front
     "f1c/0.3c+l+t",
-    # line with triangle front style, symbols are plotted on the right side of the front
-    # use other pen for the outline of the symbol
+    # line with triangle front style, symbols are plotted on the right side of
+    # the front use other pen for the outline of the symbol
     "f1c/0.4c+r+t+p1.5p,dodgerblue",
-    # line with triangle front style, symbols are plotted on the right side of the front
-    # and offset is set to 0.3 cm, skip the outline
+    # line with triangle front style, symbols are plotted on the right side of
+    # the front and offset is set to 0.3 cm, skip the outline
     "f0.5c/0.3c+r+t+o0.3c+p",
-    # line with triangle front style, symbols are plotted on the right side of the front
-    # and offset is set to 0.3 cm, skip the outline and make the main front line invisible
+    # line with triangle front style, symbols are plotted on the right side of
+    # the front and offset is set to 0.3 cm, skip the outline and make the main
+    # front line invisible
     "f0.5c/0.3c+r+t+o0.3c+p+i",
 ]:
     y -= 1  # move the current line down
