+++ examples/gallery/symbols/datetime_inputs.py
@@ -7,8 +7,9 @@ Datetime inputs of the following types are supported in PyGMT:
 - :class:`numpy.datetime64`
 - :class:`pandas.DatetimeIndex`
 - :class:`xarray.DataArray`: datetimes included in a *xarray.DataArray*
-- raw datetime strings in `ISO format <https://en.wikipedia.org/wiki/ISO_8601>`__
-  (e.g. ``"YYYY-MM-DD"``, ``"YYYY-MM-DDTHH"``, and ``"YYYY-MM-DDTHH:MM:SS"``)
+- raw datetime strings in
+  `ISO format <https://en.wikipedia.org/wiki/ISO_8601>`__ (e.g.
+  ``"YYYY-MM-DD"``, ``"YYYY-MM-DDTHH"``, and ``"YYYY-MM-DDTHH:MM:SS"``)
 - Python built-in :class:`datetime.datetime` and :class:`datetime.date`
 
 We can pass datetime inputs based on one of the types listed above directly to
