### 2021-08-31 fa9c895
Update line lengths for tutorials
