+++ examples/tutorials/coastlines.py
@@ -8,7 +8,7 @@ Plotting coastlines and borders is handled by :meth:`pygmt.Figure.coast`.
 
 import pygmt
 
-########################################################################################
+###############################################################################
 # Shorelines
 # ----------
 #
@@ -19,7 +19,7 @@ fig.basemap(region="g", projection="W15c", frame=True)
 fig.coast(shorelines=True)
 fig.show()
 
-########################################################################################
+###############################################################################
 # The shorelines are divided in 4 levels:
 #
 # 1. coastline
@@ -27,16 +27,16 @@ fig.show()
 # 3. island-in-lake shore
 # 4. lake-in-island-in-lake shore
 #
-# You can specify which level you want to plot by passing the level number and a GMT pen
-# configuration. For example, to plot just the coastlines with 0.5 thickness and black
-# lines:
+# You can specify which level you want to plot by passing the level number and
+# a GMT pen configuration. For example, to plot just the coastlines with 0.5
+# thickness and black lines:
 
 fig = pygmt.Figure()
 fig.basemap(region="g", projection="W15c", frame=True)
 fig.coast(shorelines="1/0.5p,black")
 fig.show()
 
-########################################################################################
+###############################################################################
 # You can specify multiple levels (with their own pens) by passing a list to
 # ``shorelines``:
 
@@ -45,12 +45,12 @@ fig.basemap(region="g", projection="W15c", frame=True)
 fig.coast(shorelines=["1/1p,black", "2/0.5p,red"])
 fig.show()
 
-########################################################################################
+###############################################################################
 # Resolutions
 # -----------
 #
-# The coastline database comes with 5 resolutions. The resolution drops by 80% between
-# levels:
+# The coastline database comes with 5 resolutions. The resolution drops by 80%
+# between levels:
 #
 # 1. ``"c"``: crude
 # 2. ``"l"``: low (default)
@@ -65,13 +65,13 @@ for res in ["c", "l", "i", "h", "f"]:
     fig.shift_origin(xshift="5c")
 fig.show()
 
-########################################################################################
+###############################################################################
 # Land and water
 # --------------
 #
-# Use the ``land`` and ``water`` parameters to specify a fill color for land and water
-# bodies. The colors can be given by name or hex codes (like the ones used in HTML and
-# CSS):
+# Use the ``land`` and ``water`` parameters to specify a fill color for land
+# and water bodies. The colors can be given by name or hex codes (like the ones
+# used in HTML and CSS):
 
 fig = pygmt.Figure()
 fig.basemap(region="g", projection="W15c", frame=True)
