### 2021-08-31 04ed562
Update whitespace and line length in projection gallery
