+++ examples/projections/cyl/cyl_mercator.py
@@ -2,14 +2,15 @@
 Mercator
 ========
 
-The Mercator projection takes its name from the Flemish cartographer Gheert Cremer,
-better known as Gerardus Mercator, who presented it in 1569. The projection is a
-cylindrical and conformal, with no distortion along the equator. A major navigational
-feature of the projection is that a line of constant azimuth is straight. Such a line
-is called a rhumb line or loxodrome. Thus, to sail from one point to another one only
-had to connect the points with a straight line, determine the azimuth of the line, and
-keep this constant course for the entire voyage. The Mercator projection has been used
-extensively for world maps in which the distortion towards the polar regions grows
+The Mercator projection takes its name from the Flemish cartographer Gheert
+Cremer, better known as Gerardus Mercator, who presented it in 1569. The
+projection is a cylindrical and conformal, with no distortion along the
+equator. A major navigational feature of the projection is that a line of
+constant azimuth is straight. Such a line is called a rhumb line or loxodrome.
+Thus, to sail from one point to another one only had to connect the points with
+a straight line, determine the azimuth of the line, and keep this constant
+course for the entire voyage. The Mercator projection has been used extensively
+for world maps in which the distortion towards the polar regions grows
 rather large.
 
 **m**\ [*lon0[/lat0]*]\ */scale* or **M**\ [*lon0*][*/lat0*]\ */width*
