+++ examples/projections/nongeo/cartesian_logarithmic.py
@@ -3,8 +3,8 @@ Cartesian logarithmic
 =====================
 
 **X**\ *width*\ [**l**]/[*height*\ [**l**]]: Give the *width* of the figure and
-the optional *height*. The axis or axes with a logarithmic transformation requires
-**l** after its size argument.
+the optional *height*. The axis or axes with a logarithmic transformation
+requires **l** after its size argument.
 """
 import numpy as np
 import pygmt
@@ -27,7 +27,8 @@ fig.plot(
     frame=["WSne+gbisque", "x2g3", "ya2f1g2"],
     x=xline,
     y=yline,
-    # Set the line thickness to *1p*, the color to *blue*, and the style to *dash*
+    # Set the line thickness to *1p*, the color to *blue*, and the style to
+    # *dash*
     pen="1p,blue,-",
 )
 # Plot square root values as points on the line
