+++ examples/projections/misc/misc_sinusoidal.py
@@ -2,11 +2,12 @@
 Sinusoidal
 ==========
 
-The sinusoidal projection is one of the oldest known projections, is equal-area, and
-has been used since the mid-16th century. It has also been called the
-"Equal-area Mercator" projection. The central meridian is a straight line; all other
-meridians are sinusoidal curves. Parallels are all equally spaced straight lines, with
-scale being true along all parallels (and central meridian).
+The sinusoidal projection is one of the oldest known projections, is
+equal-area, and has been used since the mid-16th century. It has also been
+called the "Equal-area Mercator" projection. The central meridian is a straight
+line; all other meridians are sinusoidal curves. Parallels are all equally
+spaced straight lines, with scale being true along all parallels (and central
+meridian).
 
 **i**\ [*lon0/*]\ *scale* or **I**\ [*lon0/*]\ *width*
 

[…]
