+++ examples/gallery/lines/linefronts.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Line fronts
 -----------
 
