+++ examples/projections/misc/misc_winkel_tripel.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Winkel Tripel
 =============
 
