+++ examples/projections/cyl/cyl_miller.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Miller cylindrical
 ==================
 
