+++ examples/projections/cyl/cyl_universal_transverse_mercator.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Universal Transverse Mercator
 =============================
 
