+++ examples/projections/cyl/cyl_stereographic.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Cylindrical Stereographic
 =========================
 
