+++ examples/projections/cyl/cyl_equidistant.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Cylindrical equidistant
 =======================
 
