+++ examples/projections/cyl/cyl_cassini.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Cassini Cylindrical
 ============================
 
