+++ examples/gallery/embellishments/colorbar.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Colorbar
 --------
 
