+++ examples/projections/cyl/cyl_equal_area.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Cylindrical equal-area
 ======================
 
