+++ examples/projections/misc/misc_eckertIV.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Eckert IV
 =========
 
