+++ examples/projections/misc/misc_sinusoidal.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Sinusoidal
 ==========
 
