+++ examples/projections/nongeo/cartesian_linear.py
@@ -1,4 +1,4 @@
-"""
+r"""
 Cartesian linear
 ================
 
