+++ examples/projections/misc/misc_mollweide.py
@@ -4,7 +4,7 @@ Mollweide
 
 This pseudo-cylindrical, equal-area projection was developed by the German
 mathematician and astronomer Karl Brandan Mollweide in 1805. Parallels are
-unequallyspaced straight lines with the meridians being equally spaced
+unequally spaced straight lines with the meridians being equally spaced
 elliptical arcs. The scale is only true along latitudes 40°44’ north and south.
 The projection is used mainly for global maps showing data distributions. It is
 occasionally referenced under the name homalographic projection.
