+++ examples/projections/misc/misc_winkel_tripel.py
@@ -2,8 +2,8 @@ r"""
 Winkel Tripel
 =============
 
-In 1921, the German mathematician Oswald Winkel a projection that was to strike
-a compromise between the properties of three elements (area, angle and
+In 1921, the German mathematician Oswald Winkel made a projection that was to
+strike a compromise between the properties of three elements (area, angle and
 distance). The German word "tripel" refers to this junction of where each of
 these elements are least distorted when plotting global maps. The projection
 was popularized when Bartholomew and Son started to use it in its
