### 2021-08-31 0bae8df
Add a test to check the incols parameter works for vector inputs
