### 2021-09-01 f19562d
Update pygmt/io.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
