+++ pygmt/io.py
@@ -9,8 +9,8 @@ def load_dataarray(filename_or_obj, **kwargs):
     Open, load into memory, and close a DataArray from a file or file-like
     object containing a single data variable.
 
-    This is a thin wrapper around :py:meth:`xarray.open_dataarray`. It differs
-    from :py:meth:`xarray.open_dataarray` in that it loads the DataArray into
+    This is a thin wrapper around :py:func:`xarray.open_dataarray`. It differs
+    from :py:func:`xarray.open_dataarray` in that it loads the DataArray into
     memory, gets GMT specific metadata about the grid via
     :py:meth:`GMTDataArrayAccessor`, closes the file, and returns the
     DataArray. In contrast, :py:func:`xarray.open_dataarray` keeps the file
