### 2021-09-01 bd38203
Update examples/gallery/symbols/datetime_inputs.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
