### 2021-09-01 dc11bb6
add required_z parameters, and if statement to handle missing z  vector
