### 2021-09-01 df1b0d5
add required_z parameters and pass it to data_kind
