### 2021-09-01 317a3dd
adding test_grd2xyz_outfile_incorrect_output_type
