+++ pygmt/tests/test_grd2xyz.py
@@ -67,3 +67,13 @@ def test_grd2xyz_no_outfile(grid):
     """
     with pytest.raises(GMTInvalidInput):
         grd2xyz(grid=grid, output_type="file")
+
+
+def test_grd2xyz_outfile_incorrect_output_type(grid):
+    """
+    Test that grd2xyz raises a warning when an outfile name is set but the
+    output_type is not set to file is set with no outfile.
+    """
+    with pytest.warns(RuntimeWarning):
+        with GMTTempFile(suffix=".xyz") as tmpfile:
+            grd2xyz(grid=grid, outfile=tmpfile.name, output_type="numpy")
