### 2021-09-01 1526f7f
add sph2grd to index.rst
