+++ doc/api/index.rst
@@ -96,6 +96,7 @@ Operations on grids:
     grdgradient
     grdsample
     grdtrack
+    sph2grd
     xyz2grd
 
 Crossover analysis with x2sys:
