### 2021-09-01 8a33395
Update pygmt/src/blockm.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
