+++ pygmt/src/blockm.py
@@ -25,7 +25,8 @@ def _blockm(block_method, table, outfile, x, y, z, **kwargs):
     Parameters
     ----------
     block_method : str
-        Name of the GMT module to call. Must be "blockmean" "blockmedian" or "blockmode".
+        Name of the GMT module to call. Must be "blockmean" "blockmedian" or
+        "blockmode".
 
     Returns
     -------
