### 2021-09-01 1b55f25
Update pygmt/src/blockm.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
