### 2021-09-01 aa041f3
Update pygmt/tests/test_surface.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
