### 2021-09-01 #43 @lwasser
hey there @weiji14 Welcome to pyopensci and please excuse the delayed response! We will review this submission and will get back to you with next steps soon. I should be able to find some folks in the geospatial space to review.
I can also  walk you through the JOSS piece. I'd encourage it because once our review is done, JOSS simply accepts  our review as  theirs so the work you have to do in order  to  get a cross-ref enabled citation  is small. but of course the decision is yours. happy to chat more if you have questions.
