### 2021-09-01 e33bf3a
Update pygmt/src/blockm.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
