### 2021-09-01 1c7df5d
Update pygmt/src/blockm.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
