### 2021-09-01 GenericMappingTools/pygmt#1479 @weiji14
Standardize parameter name for table data inputs
**Description of the desired feature**

_Originally posted by @meghanrjones in https://github.com/GenericMappingTools/pygmt/issues/1441#issuecomment-901303058_

> This is somewhat related - I am not sure if it's worth a deprecation but I think we should decide on whether the parameter for tabular input is `data` or `table` for all modules moving forward. Currently, this is the distribution:
>
> `data` parameter: surface, wiggle, velo, rose, plot3d, plot, contour
> `table` parameter: blockmedian, blockmean, histogram
> other: grdtrack (`points`), text (`textfiles`)
>
> It would also be worthwhile to set a consistent order for the positional arguments x, y, [z], data, [outfile]

**Are you willing to help implement and maintain this feature?** Discuss first
