### 2021-09-01 b137613
Add pygmt.load_dataarray function and refactor modules that return None or xarray.dataarray (#1439)
Add new function pygmt.load_datarray
Use pygmt.load_dataarrary for modules that return grids
Add documentation

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
