+++ doc/api/index.rst
@@ -107,6 +107,14 @@ Crossover analysis with x2sys:
     x2sys_init
     x2sys_cross
 
+Input/output
+------------
+
+.. autosummary::
+    :toctree: generated
+
+    load_dataarray
+
 GMT Defaults
 ------------
 
