### 2021-09-01 GenericMappingTools/pygmt#1479 @weiji14
One option to reduce the amount of deprecation notices is to keep two parameters, but with this rule:

- `data` - use for the `fig.*` plotting functions (e.g. `fig.plot`, `fig.contour`, etc)
- `table` - use for the data processing functions (e.g. `blockmean`, `histogram`, etc)

This way, we'll only need to do the renaming for `surface` ('data' -> 'table'), `grdtrack` ('points' -> 'table') and `text` ('textfiles' -> 'data'). I have some reservations with renaming `text`'s 'textfiles' to 'data' since it's a bit non-standard (and need to wait for #1121), but otherwise, what do people think?
