### 2021-09-01 GenericMappingTools/pygmt#1479 @meghanrjones
> One option to reduce the amount of deprecation notices is to keep two parameters, but with this rule:
>
>     * `data` - use for the `fig.*` plotting functions (e.g. `fig.plot`, `fig.contour`, etc)
>
>     * `table` - use for the data processing functions (e.g. `blockmean`, `histogram`, etc)
>
>
> This way, we'll only need to do the renaming for `surface` ('data' -> 'table'), `grdtrack` ('points' -> 'table') and `text` ('textfiles' -> 'data'). I have some reservations with renaming `text`'s 'textfiles' to 'data' since it's a bit non-standard (and need to wait for #1121), but otherwise, what do people think?

One complication is that many modules can be plotting or processing (e.g., `histogram` is mostly plotting but can return statistics with **-I**). My preference would be to use `data` moving forward and, if needed, deprecate `table` to `data` in `blockmean`, `blockmedian`, and `histogram`. `grdtrack` and `text` are non-standard enough for different names in my opinion.
