### 2021-09-02 16bc8ab
Let Flake8 check examples and fix warnings (#1477)
This PR enables flake8 checking of the examples directory
and fixes resulting warnings.

* Update line lengths in gallery
* Update line lengths for tutorials
* Update whitespace and line length in projection gallery
* Update rose line length
* Fix invalid escape sequences
* Fix grdsample sphinx warning
* Update examples/gallery/symbols/datetime_inputs.py
* Format grdlandmask example

Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
