+++ examples/gallery/images/grdlandmask.py
@@ -13,7 +13,8 @@ fig = pygmt.Figure()
 # Define region of interest
 region = [-65, -40, -40, -20]
 
-# Assign a value of 0 for all water masses and a value of 1 for all land masses.
+# Assign a value of 0 for all water masses and a value of 1 for all land
+# masses.
 # Use shoreline data with (l)ow resolution and set the grid spacing to
 # 5 arc minute in x and y direction.
 grid = pygmt.grdlandmask(region=region, spacing="5m", maskvalues=[0, 1], resolution="l")
