### 2021-09-02 ae65cfc
Refactor grdmix implementation to use pygmt.io.load_dataarray
