### 2021-09-03 bd6bf57
Refactor triangulate implementation to use pygmt.io.load_dataarray
