### 2021-09-03 cceb2a6
Alias binary(b), nodata(d), find(e), coltypes(f), header(h), incols(i)
