### 2021-09-03 GenericMappingTools/pygmt#1479 @weiji14
> One complication is that many modules can be plotting or processing (e.g., `histogram` is mostly plotting but can return statistics with **-I**).

Those types of dual-purpose functions (e.g. `histogram`, `rose`, `solar`) might end up having two APIs, depending on how we decide at #896. E.g. a `fig.histogram` for plotting and `pygmt.histogram` for extracting data.

> My preference would be to use `data` moving forward and, if needed, deprecate `table` to `data` in `blockmean`, `blockmedian`, and `histogram`. `grdtrack` and `text` are non-standard enough for different names in my opinion.

Ok, sounds like we need a vote. So the options are:

- :tada: Use `data` only for all functions, as per Meghan at https://github.com/GenericMappingTools/pygmt/issues/1479#issuecomment-910898788
- :rocket: Use `table` only for all functions
- :eyes: Use `data` for plotting functions, `table` for processing functions as per Wei Ji at https://github.com/GenericMappingTools/pygmt/issues/1479#issuecomment-910878607.
- :confused: Some other alternative - leave a comment below.

Place your vote on this comment. Only one vote per person :stuck_out_tongue:
