### 2021-09-03 e471f9e
Update pygmt/helpers/utils.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
