### 2021-09-03 11d98ee
Update pygmt/helpers/utils.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
