+++ pygmt/helpers/utils.py
@@ -64,7 +64,7 @@ def data_kind(data, x=None, y=None, z=None, required_z=False):
     if data is not None and (x is not None or y is not None or z is not None):
         raise GMTInvalidInput("Too much data. Use either data or x and y.")
     if data is None and (x is None or y is None):
-        raise GMTInvalidInput("Must provided both x and y.")
+        raise GMTInvalidInput("Must provide both x and y.")
     if data is None and required_z and (x is None or y is None or z is None):
         raise GMTInvalidInput("Must provide x, y, and z.")
 
