### 2021-09-03 bcfc03a
Update pygmt/helpers/utils.py
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
