### 2021-09-03 ba73d5b
checks that when required z data matrix has at least 3 columns
