+++ pygmt/tests/test_clib.py
@@ -422,9 +422,11 @@ def test_virtualfile_from_data_required_z_vectors():
             with lib.virtualfile_from_data(x=x, y=y, z=z, required_z=True):
                 print("This should have failed")
 
+
 def test_virtualfile_from_data_required_z_matrix():
     """
-    Test that function fails when needs and do not provide third z columns in a matrix.
+    Test that function fails when needs and do not provide third z columns in a
+    matrix.
     """
 
     data = np.ones((5, 2))
