### 2021-09-03 166b439
add required_z=True to _blockm
