### 2021-09-03 7675449
add required_z=True to contour
