### 2021-09-03 09c68cd
add required_z=True to plot3d
