+++ pygmt/src/plot3d.py
@@ -223,7 +223,7 @@ def plot3d(
     with Session() as lib:
         # Choose how data will be passed in to the module
         file_context = lib.virtualfile_from_data(
-            check_kind="vector", data=data, x=x, y=y, z=z, extra_arrays=extra_arrays
+            check_kind="vector", data=data, x=x, y=y, z=z, extra_arrays=extra_arrays, required_z=True
         )
 
         with file_context as fname:
