### 2021-09-03 8b3eeb2
add required_z=True to surface
