+++ pygmt/src/surface.py
@@ -78,9 +78,7 @@ def surface(x=None, y=None, z=None, data=None, **kwargs):
         - None if ``outfile`` is set (grid output will be stored in file set by
           ``outfile``)
     """
-    kind = data_kind(data, x, y, z)
-    if kind == "vectors" and z is None:
-        raise GMTInvalidInput("Must provide z with x and y.")
+    kind = data_kind(data, x, y, z, required_z=True)
 
     with GMTTempFile(suffix=".nc") as tmpfile:
         with Session() as lib:
