### 2021-09-04 8b937c7
add error for not monotonically increasing grid
