### 2021-09-04 fb5a71b
Update examples/gallery/images/grdgradient.py
Co-authored-by: Michael Grund <23025878+michaelgrund@users.noreply.github.com>
