### 2021-09-04 75ad23c
swaping places of original and grdient data
