### 2021-09-07 b5ad405
Replace spaces in arguments with octal code 040
Modifying build_arg_string function to replace blank space
characters with octal code 040, and added a doctest to
check various combinations with single and double quotes
included.
