### 2021-09-07 924f439
Remove workaround for spaces in fig.text's -F argument
