### 2021-09-07 707745f
Remove double quotes around legend label test examples
