+++ pygmt/tests/test_legend.py
@@ -52,7 +52,7 @@ def test_legend_entries():
         pen="faint",
         label="Apples",
     )
-    fig.plot(data="@Table_5_11.txt", pen="1.5p,gray", label='"My lines"')
+    fig.plot(data="@Table_5_11.txt", pen="1.5p,gray", label="My lines")
     fig.plot(data="@Table_5_11.txt", style="t0.15i", color="orange", label="Oranges")
     fig.legend(position="JTR+jTR")
 
