### 2021-09-07 03e4308
Edit test_rose_no_sectors to remove single quotes from title
