### 2021-09-07 095449a
Remove workaround for spaces in fig.psconvert prefix
Doesn't work yet, as the filename will contain the 040 octal
code, but committing to have the diff available for review.
