+++ pygmt/figure.py
@@ -192,8 +192,6 @@ class Figure:
         # Default cropping the figure to True
         if "A" not in kwargs:
             kwargs["A"] = ""
-        # allow for spaces in figure name
-        kwargs["F"] = f'"{kwargs.get("F")}"' if kwargs.get("F") else None
         with Session() as lib:
             lib.call_module("psconvert", build_arg_string(kwargs))
 
