### 2021-09-07 d81b80b
Format using blackdoc and fix flake8 error
