### 2021-09-07 f786ac7
rais error only when grin increment is zero
