### 2021-09-07 5b2ae22
add tests for zero increment grid
