+++ pygmt/tests/test_clib.py
@@ -742,6 +742,22 @@ def test_dataarray_to_matrix_inc_fails():
     with pytest.raises(GMTInvalidInput):
         dataarray_to_matrix(grid)
 
+def test_dataarray_to_matrix_zero_inc_fails():
+    """
+    Check that dataarray_to_matrix fails for zero increments grid.
+    """
+    data = np.ones((5, 5), dtype="float64")
+    x = np.linspace(0, 1, 5)
+    y = np.zeros_like(x)
+    grid = xr.DataArray(data, coords=[("y", y), ("x", x)])
+    with pytest.raises(GMTInvalidInput):
+        dataarray_to_matrix(grid)
+    y = np.linspace(0, 1, 5)
+    x = np.zeros_like(x)
+    grid = xr.DataArray(data, coords=[("y", y), ("x", x)])
+    with pytest.raises(GMTInvalidInput):
+        dataarray_to_matrix(grid)
+
 
 def test_get_default():
     """
