+++ pygmt/clib/conversion.py
@@ -102,9 +102,7 @@ def dataarray_to_matrix(grid):
             )
         if coord_inc == 0:
             raise GMTInvalidInput(
-                "Grid has zero increment in the '{}' dimension.".format(
-                    dim
-                )
+                "Grid has zero increment in the '{}' dimension.".format(dim)
             )
         region.extend(
             [
