### 2021-09-07 cef091b
change test format to avoid using grdinfo
