### 2021-09-07 71418f0
refactor test_sphinterpolate.py to remove grdinfo
