+++ examples/gallery/maps/roads.py
@@ -1,8 +1,12 @@
 """
 Roads
 -----
-The :meth:`pygmt.Figure.plot` method allows to plot geographical data which is
-stored in a :class:`geopandas.GeoDataFrame` object.
+The :meth:`pygmt.Figure.plot` method allows us to plot geographical data such
+as lines which are stored in a :class:`geopandas.GeoDataFrame` object. Use
+:func:`geopandas.read_file` to load data from any supported OGR format such as
+a shapefile (.shp), GeoJSON (.geojson), geopackage (.gpkg), etc. Then, pass the
+:class:`geopandas.GeoDataFrame` as an argument to the `data` parameter in
+:meth:`pygmt.Figure.plot`, and style the geometry using the `pen` parameter.
 """
 
 import geopandas as gpd
@@ -23,11 +27,8 @@ fig = pygmt.Figure()
 # Define target region around O'ahu (Hawai'i)
 region = [-158.3, -157.6, 21.2, 21.75]  # minx, maxx, miny, maxy
 
-fig.basemap(
-    region=region,
-    projection="M12c",
-    frame=["af", 'WSne+t"Main roads of Oahu (Hawaii)"'],
-)
+title = r"Main roads of O\047ahu (Hawai\047i)"  # \047 is octal code for '
+fig.basemap(region=region, projection="M12c", frame=["af", f'WSne+t"{title}"'])
 fig.coast(land="gray", water="dodgerblue4", shorelines="1p,black")
 
 # Plot the individual road types with different pen settings and assgin labels
