### 2021-09-08 cedcda2
Don't use which in test
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
