### 2021-09-08 a4bb96e
Name test more appropriately
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
