### 2021-09-08 fc43c85
Add newer common options
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
