### 2021-09-08 57aba71
Update pygmt/src/nearneighbor.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
