### 2021-09-09 089cb5c
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
