+++ examples/gallery/maps/roads.py
@@ -5,8 +5,8 @@ The :meth:`pygmt.Figure.plot` method allows us to plot geographical data such
 as lines which are stored in a :class:`geopandas.GeoDataFrame` object. Use
 :func:`geopandas.read_file` to load data from any supported OGR format such as
 a shapefile (.shp), GeoJSON (.geojson), geopackage (.gpkg), etc. Then, pass the
-:class:`geopandas.GeoDataFrame` as an argument to the `data` parameter in
-:meth:`pygmt.Figure.plot`, and style the geometry using the `pen` parameter.
+:class:`geopandas.GeoDataFrame` as an argument to the ``data`` parameter in
+:meth:`pygmt.Figure.plot`, and style the geometry using the ``pen`` parameter.
 """
 
 import geopandas as gpd
