### 2021-09-09 b247dc4
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
