### 2021-09-09 5f35b08
Apply suggestions from code review
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
