### 2021-09-09 50e0382
Add xarray functions to test outgrid info
Remove the use of grdinfo and use xarray min, max, median, and mean functions to test data in outgrid
