### 2021-09-09 3c5b4e9
removed file from maps category
