### 2021-09-09 b4f0284
Add rtol parameter for npt assertions
