### 2021-09-10 0f1f7ad
remove unused import in test_sph2grd.py
