### 2021-09-10 2edaa63
Wrap sphdistance (#1383)
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
