+++ pygmt/__init__.py
@@ -46,6 +46,7 @@ from pygmt.src import (
     grdtrack,
     info,
     makecpt,
+    sphdistance,
     surface,
     which,
     x2sys_cross,
