### 2021-09-10 cfb9572
Apply suggestions from code review
Co-authored-by: Will Schlitzer <schlitzer90@gmail.com>
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
