### 2021-09-10 de206a3
Add table docstring to parameters list
