### 2021-09-11 974f908
add table description to docstring
