### 2021-09-11 97309b0
move sphdistance to tabular data section in index
