### 2021-09-11 cd19877
add table description to docstring
