### 2021-09-11 41400f9
Apply suggestions from code review
Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>
