### 2021-09-11 6cbb18b
add dimfilter
