### 2021-09-11 c9affe3
add docstring for sectors, filters, and distance
