### 2021-09-11 GenericMappingTools/pygmt#1479 @meghanrjones
Can we also set the conventional order for the positional arguments as `data/table, x, y, [z]` or `x, y, [z], data/table`? My opinion is that this isn't worth deprecating any current syntax, but that https://github.com/GenericMappingTools/pygmt/pull/1434, https://github.com/GenericMappingTools/pygmt/pull/1429, https://github.com/GenericMappingTools/pygmt/pull/1418, https://github.com/GenericMappingTools/pygmt/pull/1379, https://github.com/GenericMappingTools/pygmt/pull/731 and https://github.com/GenericMappingTools/pygmt/pull/1491 should have the same order (if both `data/table` and `x/y[/z]` will be supported).

Here's the current count:
- `x, y, [z], data/table` - (5 current functions) contour, plot3d, wiggle, surface, plot (Used in PRs https://github.com/GenericMappingTools/pygmt/pull/1379
- `data/table, [outfile], x, y, [z]` - (2 current functions) blockmean, blockmedian (Used in PRs https://github.com/GenericMappingTools/pygmt/pull/731, https://github.com/GenericMappingTools/pygmt/pull/1456)
- only `data/table` with no `x/y[/z]` option - (2 current functions and 2 functions in v0.5.0) histogram, sphdistance, xyz2grd, info (Used in PRs https://github.com/GenericMappingTools/pygmt/pull/1429, https://github.com/GenericMappingTools/pygmt/pull/1434, https://github.com/GenericMappingTools/pygmt/pull/1418, https://github.com/GenericMappingTools/pygmt/pull/1491)
