### 2021-09-11 3373066
Add check for required arguments
