### 2021-09-11 e3a4f7a
Add test_dimfilter.py
