### 2021-09-12 9f8f769
Update pygmt/clib/conversion.py
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
