### 2021-09-12 f134c81
Update pygmt/tests/test_clib.py
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
