+++ pygmt/tests/test_clib.py
@@ -753,6 +753,7 @@ def test_dataarray_to_matrix_zero_inc_fails():
     grid = xr.DataArray(data, coords=[("y", y), ("x", x)])
     with pytest.raises(GMTInvalidInput):
         dataarray_to_matrix(grid)
+
     y = np.linspace(0, 1, 5)
     x = np.zeros_like(x)
     grid = xr.DataArray(data, coords=[("y", y), ("x", x)])
