+++ pygmt/clib/conversion.py
@@ -82,7 +82,7 @@ def dataarray_to_matrix(grid):
     """
     if len(grid.dims) != 2:
         raise GMTInvalidInput(
-            "Invalid number of grid dimensions '{}'. Must be 2.".format(len(grid.dims))
+            f"Invalid number of grid dimensions '{len(grid.dims)}'. Must be 2."
         )
     # Extract region and inc from the grid
     region = []
@@ -96,9 +96,7 @@ def dataarray_to_matrix(grid):
         coord_inc = coord_incs[0]
         if not np.allclose(coord_incs, coord_inc):
             raise GMTInvalidInput(
-                "Grid appears to have irregular spacing in the '{}' dimension.".format(
-                    dim
-                )
+                f"Grid appears to have irregular spacing in the '{dim}' dimension."
             )
         if coord_inc == 0:
             raise GMTInvalidInput(
