### 2021-09-12 8451465
remove insuficent data test from test_contour
