### 2021-09-12 6a0042e
remove test for passing z value test_surface
