+++ pygmt/tests/test_contour.py
@@ -3,12 +3,10 @@
 Tests contour.
 """
 import os
-from itertools import product
 
 import numpy as np
 import pytest
 from pygmt import Figure
-from pygmt.exceptions import GMTInvalidInput
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
 POINTS_DATA = os.path.join(TEST_DATA_DIR, "points.txt")
