+++ .github/workflows/ci_tests.yaml
@@ -146,6 +146,6 @@ jobs:
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2.1.0
         with:
-          files: ./coverage.xml # optional
+          file: ./coverage.xml # optional
           env_vars: OS,PYTHON,NUMPY
           fail_ci_if_error: false
