### 2021-09-12 0cfd82b
Revert "Add tests for load_dataarray to increase code coverage"
This reverts commit 6bbc77926a63e0e11e35da47d763be4ece8ae27b.
