### 2021-09-12 19015ab
Drop deprecated codecov dependency from conda install
The codecov-python package is being deprecated, see
https://github.com/codecov/codecov-python/pull/325.
