+++ .github/workflows/ci_tests.yaml
@@ -92,7 +92,7 @@ jobs:
           conda install gmt=6.2.0 numpy=${{ matrix.numpy-version }} \
                         pandas xarray netCDF4 packaging \
                         ${{ matrix.optional-packages }} \
-                        codecov coverage[toml] dvc ipython make \
+                        coverage[toml] dvc ipython make \
                         pytest-cov pytest-mpl pytest>=6.0 \
                         sphinx-gallery
 
