### 2021-09-15 GenericMappingTools/pygmt#1479 @weiji14
> Can we also set the conventional order for the positional arguments as `data/table, x, y, [z]` or `x, y, [z], data/table`?

Yes, good point. I would say just go with `data, x, y, [z]`, so people can do `pygmt.somefunction(data)` if `data` is a `pandas.DataFrame`/`geopandas.GeoDataFrame` or any table-like input, similar to other Python-like plotting/processing code like `plt.imshow`. In this case, we would need to modify:

> * `x, y, [z], data/table` - (5 current functions) contour, plot3d, wiggle, surface, plot (Used in PRs [Wrap nearneighbor #1379](https://github.com/GenericMappingTools/pygmt/pull/1379)

Debating on whether we should go through a deprecation cycle (2 releases) for the `x/y/z/data` -> `data/x/y/z` parameter ordering change. Personally I try to always do `parameter=arg` in my code, but many people just pass in the arguments directly. So maybe best to raise a warning to give people time to switch over :upside_down_face:
