### 2021-09-22 GenericMappingTools/pygmt#1479 @seisman
Just want to confirm that we have reached a consensus that we will use `data` and the order `data, x, y, z` for table inputs. If yes, then we can start to update these functions.

> If we will make the change for the functions that have `x/y/z/data`, I think we should raise a warning.

It's unclear to me how to raise a warning for this case. For users who do `x=x, y=y, z=z`, the parameter order doesn't matter, thus the warning is unnecessary and may be confusing. For users who do `x, y, z`, how can know they pass arguments directly without giving parameters?
