### 2021-09-23 GenericMappingTools/pygmt#1479 @seisman
Here are a list of modules that may need to rename `table` to `data` or reorder parameters like `data, x, y, z`:

- [ ] blockm.py: `blockm(block_method, table, outfile, x, y, z, **kwargs)`
- [ ] blockm.py: `blockmean(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs)`
- [ ] blockm.py: `blockmedian(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs)`
- [ ] blockm.py: `blockmode(table=None, outfile=None, *, x=None, y=None, z=None, **kwargs)`
- [ ] contour.py: `contour(self, x=None, y=None, z=None, data=None, **kwargs)`
- [ ] grdtrack.py: `grdtrack(points, grid, newcolname=None, outfile=None, **kwargs)`
- [ ] histogram.py: `histogram(self, table, **kwargs)`
- [ ] info.py: `info(table, **kwargs)`
- [ ] plot.py: `plot(self, x=None, y=None, data=None, size=None, direction=None, **kwargs)`
- [ ] plot3d.py: `plot3d(self, x=None, y=None, z=None, data=None, size=None, direction=None, **kwargs)
- [ ] rose.py: `rose(self, length=None, azimuth=None, data=None, **kwargs)`
- [ ] sphdistance.py: `sphdistance(table, **kwargs)`
- [ ] surface.py: `surface(x=None, y=None, z=None, data=None, **kwargs)`
- [ ] text.py: `text_(self, textfiles=None, x=None, y=None, position=None, text=None, angle=None, font=None, justify=None, **kwargs)
- [ ] velo.py: `velo(self, data=None, **kwargs)`
- [ ] wiggle.py: `wiggle(self, x=None, y=None, z=None, data=None, **kwargs)`
- [ ] x2sys_cross.py: `x2sys_cross(tracks=None, outfile=None, **kwargs)`
- [ ] xyz2grd.py: `xyz2grd(table, **kwargs)`
