### 2021-09-24 GenericMappingTools/pygmt#1479 @seisman
> If the user passes in `x, y, z`, then `len(args)` will be at least 3. If the user passes in `x=x, y=y, z=y`, then `len(args)` will be 0, using a decorator similar to `deprecate_parameter`.

It looks like a good solution. I'll work on it.
