### 2021-09-27 95fa173
Fix the check_data_input_order decorator
