### 2021-09-27 GenericMappingTools/pygmt#1479 @seisman
> If the user passes in `x, y, z`, then `len(args)` will be at least 3. If the user passes in `x=x, y=y, z=y`, then `len(args)` will be 0, using a decorator similar to `deprecate_parameter`.

I'm rethinking the decorator implemented in #1541.

We're changing the parameter order from `x, y, [z], data` to `data, x, y, [z]`. Before making the change, users can pass positional parameters `x, y, **kwargs` or `x, y, z, **kwargs`. After making the change, users still can pass positional parameters, but they can only use it in this way `data, **kwargs`.

Thus, actually, we should raise the warning only if `len(args) > 1` (i.e., `x, y` or `x, y, z`), for non-plotting functions.

For plotting functions, there is always the `self` parameter, thus, we should check if `len(args) > 2` instead.
