### 2021-09-27 GenericMappingTools/pygmt#1479 @seisman
> For example, in #1546 someone could pass in `data, length, azimuth` as positional arguments and should not get a warning even though `len(args>2)`.

Users can pass `data`, `data=data`, or `length=length, azimuth=azimuth`, but they can't do `data, length, azimuth` because we have the `data_kind` function to make sure that users cannot pass both `data` and `x,y`.
