### 2021-09-28 GenericMappingTools/pygmt#1479 @seisman
> We're changing the parameter order from `x, y, [z], data` to `data, x, y, [z]`. Before making the change, users can pass positional parameters `x, y, **kwargs` or `x, y, z, **kwargs`. After making the change, users still can pass positional parameters, but they can only use it in this way `data, **kwargs`.
>
> Thus, actually, we should raise the warning only if `len(args) > 1` (i.e., `x, y` or `x, y, z`), for non-plotting functions.
>
> For plotting functions, there is always the `self` parameter, thus, we should check if `len(args) > 2` instead.

I've implemented it in 95fa173cbba251096ab0f22764cbfb11824cd07b as part of PR #1548.

> > > For example, in #1546 someone could pass in `data, length, azimuth` as positional arguments and should not get a warning even though `len(args>2)`.
> >
> >
> > Users can pass `data`, `data=data`, or `length=length, azimuth=azimuth`, but they can't do `data, length, azimuth` because we have the `data_kind` function to make sure that users cannot pass both `data` and `x,y`.
>
> OK, that was a bad example. An alternate example for plot: `data, None, None, size`. I don't know why anyone would want to do this, but they could.

I prefer NOT to check these complicated cases:

1. Although users can pass `data, None, None, size`, I believe it's very rare. The current decorator only raise a warning, so users script won't stop working.
2. The decorator is temporary, and likely will be removed after v0.7.0. It's six months if we strictly following the 3-month release cycle.
