### 2021-10-03 GenericMappingTools/pygmt#1479 @seisman
After #1565 is merged, all functions will use consistent parameter name `data`, and consistent parameter order `data, x, y, [z]`. There are few exceptions (see https://www.pygmt.org/dev/api/index.html):

- `image(self, imagefile, ...)`
- `legend(self, spec=None, ...)`
- `meca(self, spec, scale, ...)`
- `rose(data=None, length=None, azimuth=None, ...)`
- `text(textfiles=None, x=None, y=None, ...)`
- `grdtrack(points, grid, ...)`
- `x2sys_cross(tracks=None, outfile=None, ...)`

I think these exceptions are reasonable and don't need more changes. Thus, we can close the issue after #1565 is merged. Agree?
