### 2021-11-24 #43 @SimonMolinsky
@lwasser, what are the next steps? Now I have a window to review the package. I assume that December will be rather busy so it's better to start now :)
