### 2021-11-24 #43 @weiji14
Yep, we're ok with people making issues on the repo. The 4 week timing sounds good too.
