### 2021-12-14 GenericMappingTools/pygmt#1662 @jbusecke
Quickinstall in the README?
**Description of the desired feature**

I really appreciate the detailed README, but I was wondering if a one liner of the recommended install method, e.g. `conda install -c conda-forge pygmt` would be a useful addition somewhere in the upper part of the file?

**Are you willing to help implement and maintain this feature?** No (just part of https://github.com/pyOpenSci/software-review/issues/43)
