### 2021-12-14 GenericMappingTools/pygmt#1663 @jbusecke
Add installing with mamba to docs?
**Description of the desired feature**
I just tried to install pygmt on the pangeo cloud deployment with conda, and it took quite long. I actually cancelled the install and used mamba instead. It was leaps and bounds faster.

I think it would really improve the user experience if you could add `mamba install -c conda-forge pygmt` (maybe with a short link to the mamba docs) to the install docs.
