### 2021-12-14 GenericMappingTools/pygmt#1664 @jbusecke
Note about related projects
I really appreciate the inclusion of [related projects](https://github.com/GenericMappingTools/pygmt#related-projects) in the README. But it might be worth mentioning that most (all?) of these are not actively maintained anymore. That would have probably saved me a few clicks and reads.
