### 2021-12-14 GenericMappingTools/pygmt#1665 @jbusecke
Ah I just saw this: `At least one of the parameters frame, map_scale, rose or compass must be specified.` [here](https://www.pygmt.org/latest/api/generated/pygmt.Figure.basemap.html#pygmt-figure-basemap).

I find this signature very confusing. Is there a way to provide a reasonable default (like `frame=True`), so that a call to `fig.basemap()` does not raise an error? I might also misunderstand something here.
