### 2021-12-14 GenericMappingTools/pygmt#1666 @jbusecke
Missing Examples for user facing function
**Description of the desired feature**

First of all, the existing examples are really, really nice! They helped me a lot understanding the functionality.

I was however not able to find examples for `pygmt.Figure.psconvert`, `pygmt.Figure.grd2cpt`, and `pygmt.set_display`.

Furthermore many of the [processing functions](https://www.pygmt.org/latest/api/index.html#data-processing) are not show in the examples as far as I can tell.
