### 2021-12-14 GenericMappingTools/pygmt#1665 @weiji14
@willschlitzer, do you want to have a think about the `frame` parameter here since you're working on revamping the first tutorial at #1607?
