### 2021-12-14 #43 @jbusecke
- https://github.com/GenericMappingTools/pygmt/issues/1664
- [x] **Installation instructions:** for the development version of package and any non-standard dependencies in README.
    - I noted [here](https://github.com/GenericMappingTools/pygmt/issues/1663) that installing with mamba is much faster and perhaps could be added to the install docs.
    - Some issues with the method signature when doing slight modifications [example](https://github.com/GenericMappingTools/pygmt/issues/1665) (for more see comments below)
- [x] **Function Documentation:** for all user-facing functions
    - See below.
    > BTW this seems somewhat redundant with the point in 'Useability' below? Or is there a difference I did overlook @lwasser
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
    - Suggested adding a [conda badge](https://github.com/GenericMappingTools/pygmt/pull/1661)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
    - https://github.com/GenericMappingTools/pygmt/issues/1662)
- [x] ~~Any additional setup required (authentication tokens, etc)~~
- [x] Brief demonstration usage
	- Not quite "brief", but the video demo is awesome!
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages
    - https://github.com/GenericMappingTools/pygmt/issues/1664)

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand
    - Most but not all. See https://github.com/GenericMappingTools/pygmt/issues/1666

- ~~[ ] **Performance:** Any performance claims of the software been confirmed.~~
    - (I didnt find any particular statements about performance and my examples generally ran zippy.
- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

Estimated hours spent reviewing: 3-4

---

My general assesment of pygmt is very positive. The code seems very well organized, and the documentation is overall excellent (the examples really helped me understand the capabilities of GMT and pygmt)! So overall I would happily recommend this package to be accepted. I gave a few minor suggestions (see above), but none of these is really a 'major' revision from my end.

The one major friction point to me was the (I guess GMT native?) syntax for e.g. setting titles. In several places I was wondering if it wouldnt be possible to (optionally) hide this aspect of GMT even more from the user? Most of the other syntax is very easy to understand for someone like me (who uses matplotlib very often), but the whole time I wanted a `fig.set_title()` or `fig.title()` or even `fig.basemap(frame='a', title='...') ` instead of `fig.basemap(frame=["a", '+t"Trinidad and Tobago"'])`. Providing these rather complex string lists makes this package much harder to use in my eyes. I do not at all regard this as an issue that needs to be implemented for the review, but I still wanted to point this out.

Thanks to all the maintainers for this great work!

A couple of random comments:

- You say "GMT shines when it comes to plotting data on a map." Should the [first figure section](https://www.pygmt.org/latest/tutorials/first_figure.html#making-your-first-figure) of the docs maybe include a simple arrow from one location to another to show this off?
