### 2021-12-20 GenericMappingTools/pygmt#1665 @willschlitzer
> @willschlitzer, do you want to have a think about the `frame` parameter here since you're working on revamping the first tutorial at #1607?

I like the idea of setting `frame=True` as the default for `basemap`, as I think that is much more likely to be plotted than `map_scale`, `rose`, or `compass` when plotting `basemap`.

Admittedly, I also think `basemap` should be rarely used when creating plots, as frames and titles can also be set in functions like `coast`, `plot`, `grdimage`, and the like. I chose not to include it on the new `first_figure.py` because I think it adds extra complexity when an adequate map can be created using only `coast`.
