### 2021-12-20 GenericMappingTools/pygmt#1662 @willschlitzer
I think this is a good idea.
