### 2021-12-20 GenericMappingTools/pygmt#1666 @willschlitzer
> **Action point**: I think inline docstring might be the way to go with some of these examples. Upstream GMT has done this (e.g. at https://docs.generic-mapping-tools.org/6.3/basemap.html#examples), so we can probably translate these to PyGMT. The [`pygmt.grdfilter`](https://www.pygmt.org/v0.5.0/api/generated/pygmt.grdfilter.html) function has an example we can use as a reference. Key challenge might be to prevent these inline code examples from running in CI (but definitely check that the syntax is ok).

I agree with using inline examples. I use the GMT inline examples to demonstrate the syntax for non-plotting modules when writing the tests for module-wrapping.
