### 2021-12-21 GenericMappingTools/pygmt#1666 @willschlitzer
@GenericMappingTools/pygmt-maintainers Would it be a good idea to add inline docstring examples (without an associated plot) to all of our module docstrings? That way the page would be a one-stop-shop without needing to go to an example.
