### 2021-12-24 GenericMappingTools/pygmt#1686 @willschlitzer
Add inline code examples in module docstrings
As first discussed in [this comment](https://github.com/GenericMappingTools/pygmt/issues/1666#issuecomment-993962742), should module docstrings contain inline code examples? It wouldn't create a plot on the module page, but would show a working example of the module with the correct syntax without needing to find an example page

**Are you willing to help implement and maintain this feature?** Yes/No
