### 2021-12-24 #43 @lwasser
hi everyone! @jbusecke thank you for your review. @SimonMolinsky we look forward to your review when you are ready to submit. Do you have any questions or need any help?

The next week is a holiday in the US and we are completely off starting this afternoon  (december 24) through the new year. I will be back online on January 3 2022!

In the meantime, @SimonMolinsky if you can wrap up your review in the next week that would be wonderful. @weiji14 feel free to begin working on the reviewer comments and and addressing them. Please reference this issue as you submit pr's and open issues to address so we have some documentation of how the package was improved / modified through this process.

All - happy early new year to you! If you need anything today please reach out. otherwise i will check back in - in January 2022 and we can work on wrapping up this review!
