### 2022-01-02 #43 @SimonMolinsky
- [  ] **Installation instructions:** for the development version of package and any non-standard dependencies in README

Not provided, but @jbusecke signaled it: [here](https://github.com/GenericMappingTools/pygmt/issues/1662)

Not provided in a textual form but linked as a tutorial video, so for me, it's ok. However, I think that the short copy & paste example below the clip could be applicable for more advanced users who would like to test the package quickly. For example, it could look like:

- Youtube lecture,
- Short installation,
- Copy/Paste example.

It was stated by the @jbusecke : [here](https://github.com/GenericMappingTools/pygmt/issues/1666). But for me explanations in the Issue are clear and I accept this point.

- [x] **Examples** for all user-facing functions

The package has a vast number of examples and tutorials.

- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [  ] Installation instructions
- [  ] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [  ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

It is not compared to competing packages in the README. I think it will be a great addition for the future because you'll be forced to do it anyway for journal articles.

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

I've installed the package on the macOS (Intel setup) without any problems.

I've run all the example tutorials with my datasets and everything has worked fine, but I have a few suggestions (at the bottom of a review).

The package is fast and the download of package datasets runs smoothly.

- [  ] **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

Yes, the test coverage is 99% and **pyGMT** has a description of tests in Contributing Guide.  But I have two comments here:

1. For full test coverage you need to install not only `pytest` but `pytest-mpl` and `sphinx-gallery` too. Those information are presented in the documentation [here](https://www.pygmt.org/dev/contributing.html#testing-your-code) but in different paragraphs. Maybe it is a good idea to rewrite this part of documentation slightly and present list of all additional packages needed for testing?
2. The other problem is more confusing. When I run all tests then I got results with 2/3 of tests passing but there is 1/3 that fail. When I look into the test output I see a text `This is expected for new tests`. So I run test again and I see the same communicate. It is extremely misleading, especially for contributors - if I create new functionalities and run all tests to check if everything is ok I must manually check each failed test if it has phrase `This is expected for new tests`. Did I miss something?

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

Yes, the package has Github Actions defined for testing and CI for the leading Operating Systems: Ubuntu (Linux), macOS, Windows.

- [  ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

---

The general feel is very good! I really liked the package but more important is ecosystem around it. The great resources and documentation makes use of the **pyGMT** very simple and I can create desired maps very fast. I'll definitively move from the old matplotlib into **pyGMT**.

However, I have few comments about the usability that are not falling into issues, rather suggestions or recommendations:

1. I don't understand why examples in the package repository are presented as Python files. Why not Notebooks? Is this related to the automatic document conversion? You can transform Notebooks with [nbconvert](https://nbconvert.readthedocs.io/en/latest/) into the documentation. I rather use Github than package website to explore its capabilities and it will be much easier to follow stylized Notebook. I'm not a cartographer so there is additional overhead to understand what's going on within the script and explanations written as a Python comments are very unfriendly for a reader. And for cartographers the images within examples may be more useful than the code itself. If there is a big block of text then I start to get lost within it. **If you are into the change I can create those notebooks and check if documentation converts appropriately** because I understand that this take time. On the other hand, presentation as a script forces user to code, not only download notebook and run cells. For education it's a good thing. But from the perspective of Data Scientist that need your package for work it complicates matter a lot.
2. There are cryptic errors which are coming from **GMT**. The example - I'll change borders of the region bounding box that lower coordinate comes after the higher:

# This is ok
fig.coast(region=[20, 31, 59, 71], shorelines=True)

# This isn't
fig.coast(region=[31, 20, 59, 71], shorelines=True)
```

And error:

```
GMTCLibError                              Traceback (most recent call last)
/var/folders/72/m9ty2jns5m96fc2_2ll55vyh0000gn/T/ipykernel_45878/1371627446.py in <module>
----> 1 fig.coast(region=[31, 20, 59, 71], shorelines=True)

~/miniconda3/envs/pygmt/lib/python3.10/site-packages/pygmt/helpers/decorators.py in new_module(*args, **kwargs)
    584                     )
    585                     warnings.warn(msg, category=SyntaxWarning, stacklevel=2)
--> 586             return module_func(*args, **kwargs)
    587
    588         new_module.aliases = aliases

~/miniconda3/envs/pygmt/lib/python3.10/site-packages/pygmt/helpers/decorators.py in new_module(*args, **kwargs)
    724                         kwargs[arg] = separators[fmt].join(f"{item}" for item in value)
    725             # Execute the original function and return its output
--> 726             return module_func(*args, **kwargs)
    727
    728         return new_module

~/miniconda3/envs/pygmt/lib/python3.10/site-packages/pygmt/src/coast.py in coast(self, **kwargs)
    190         )
    191     with Session() as lib:
--> 192         lib.call_module("coast", build_arg_string(kwargs))

~/miniconda3/envs/pygmt/lib/python3.10/site-packages/pygmt/clib/session.py in call_module(self, module, args)
    498         )
    499         if status != 0:
--> 500             raise GMTCLibError(
    501                 f"Module '{module}' failed with status code {status}:\n{self._error_message}"
    502             )

GMTCLibError: Module 'coast' failed with status code 72:
coast [ERROR]: Option -R parsing failure. Correct syntax:
coast [ERROR]: Offending option -R31/20/59/71
```

I understand that it is a lot of work to catch those exceptions and make them more readable. At the same time I highly recommend to do so. From my perspective, errors and exceptions from the wrapped C/C++ libraries sometimes may generate more noise than help, and it may affect the decision if package will be used or not. (The good example are GDAL errors, or TensorFlow errors - we are not happy if we need to set software based on those packages in production). This could be done because:

3. The code needs to be more Pythonic. If I'm forced to pass strings like that `shorelines='1/0.5p,gray' I would rather go for the **GMT** - why to bother with Python if syntax is not Pythonic? I understand that it is a process under way and I saw your development stages and new aliases build upon some parameters and functions and I really appreciate that. But I'd like to emphasize that this syntax should be available along with Python dictionaries to set image parameters so don't leave loose ends. String syntax make code automation and debugging much harder, and I even don't talk about the readability.

Anyway, thanks a lot! I'm into **pyGMT** and I love your work and time put into the package.
