### 2022-01-07 #43 @leouieda
Happy New Year everyone! 🎊  Thank you for the reviews and very positive comments!

Just wanted to chime in on the use of notebooks for examples since I set this up way back when we first started the package:

* We use sphinx-gallery to generate both the gallery examples and the tutorials from the Python files. They generate the notebooks automatically when building the docs (see links at the end of each tutorial page).
* Notebooks are great for users but reviewing changes to them in Pull Requests is a nightmare for maintainers since diffs are unreadable without using some third party tool like ReviewNB.
* Notebooks that have images can get very large, which causes active repositories to quickly become huge (~100s-1000s Mb).
* It's relatively easy to fix a typo or make a small change in a script, less so in a notebook (need to open jupyter locally).
* Enforcing style and formatting (flake8/black/etc) in notebooks is not trivial.

I had initially tried using notebooks in previous projects but always came to regret it after a while because of these problems.
