### 2022-01-07 #43 @lwasser
hey there @leouieda  👋 !!
I think that all of your comments above are reasonable. Sphinx gallery is great because it does provide that download link in both notebook and .py format at the bottom of each gallery item so i think it addresses the issue of anyone who might want a notebook without the issues you mentioned above. Thank you so much for the thoughtful reply. I've also thought about building a documentation working group so we can have more information about options for documentation such as sphinx gallery and some of the potential pitfalls such as those you mentioned above!

When you and the pygmt team have time, please give us a general sense of how much effort those review comments will take to work through. As with any review, you are fully welcome to respond to comments and address issues as the team sees fit.

Thank you again @jbusecke and @SimonMolinsky for your reviews. Once the pygmt team has addressed both reviews, I will ask you both to address whether the package updates address the items that you highlight sufficiently.
