### 2022-01-08 #43 @weiji14
Hi all, thanks for all the great feedback. I will provide a separate response to all the longer review comments, but the ones below seem like the key issues the PyGMT team will need to work on in the coming weeks:

- [ ] https://github.com/GenericMappingTools/pygmt/issues/1662
- [ ] https://github.com/GenericMappingTools/pygmt/issues/1664
- [ ] https://github.com/GenericMappingTools/pygmt/issues/1697

My estimate that we can complete these in 1-2 weeks, or by then end of January at the very latest. Let me know if I missed anything.
