### 2022-01-08 #43 @weiji14
### Responses to Review Comments

> @jbusecke - The one major friction point to me was the (I guess GMT native?) syntax for e.g. setting titles. In several places I was wondering if it wouldnt be possible to (optionally) hide this aspect of GMT even more from the user? Most of the other syntax is very easy to understand for someone like me (who uses matplotlib very often), but the whole time I wanted a `fig.set_title()` or `fig.title()` or even `fig.basemap(frame='a', title='...') ` instead of `fig.basemap(frame=["a", '+t"Trinidad and Tobago"'])`. Providing these rather complex string lists makes this package much harder to use in my eyes. I do not at all regard this as an issue that needs to be implemented for the review, but I still wanted to point this out.

> @SimonMolinsky - 3. The code needs to be more Pythonic. If I'm forced to pass strings like that `shorelines='1/0.5p,gray' I would rather go for the **GMT** - why to bother with Python if syntax is not Pythonic? I understand that it is a process under way and I saw your development stages and new aliases build upon some parameters and functions and I really appreciate that. But I'd like to emphasize that this syntax should be available along with Python dictionaries to set image parameters so don't leave loose ends. String syntax make code automation and debugging much harder, and I even don't talk about the readability.

Having Pythonic GMT arguments is something that has came up time and time again, and is a long term goal for PyGMT (see https://github.com/GenericMappingTools/pygmt/projects/6) we want to get done by v1.0. There is some debate on how the syntax should look like (see https://github.com/GenericMappingTools/pygmt/issues/1082), and it will take time to fully implement all the possible GMT shortcodes. At this stage, we need to balance our developer resources on adding new features (functionality) and better API design (usability), but it is good to hear from both reviewers that the UX aspect is something we need to spend more time on! :grin:

> A couple of random comments:
>
> * You say "GMT shines when it comes to plotting data on a map." Should the [first figure section](https://www.pygmt.org/latest/tutorials/first_figure.html#making-your-first-figure) of the docs maybe include a simple arrow from one location to another to show this off?

There's currently work going on to revamp the first figure tutorial (see https://github.com/GenericMappingTools/pygmt/issues/770). Ping @willschlitzer so that he's aware of this suggestion :)

> @SimonMolinsky - 1. I don't understand why examples in the package repository are presented as Python files. Why not Notebooks? Is this related to the automatic document conversion? You can transform Notebooks with [nbconvert](https://nbconvert.readthedocs.io/en/latest/) into the documentation. I rather use Github than package website to explore its capabilities and it will be much easier to follow stylized Notebook. I'm not a cartographer so there is additional overhead to understand what's going on within the script and explanations written as a Python comments are very unfriendly for a reader. And for cartographers the images within examples may be more useful than the code itself. If there is a big block of text then I start to get lost within it. **If you are into the change I can create those notebooks and check if documentation converts appropriately** because I understand that this take time. On the other hand, presentation as a script forces user to code, not only download notebook and run cells. For education it's a good thing. But from the perspective of Data Scientist that need your package for work it complicates matter a lot.

I don't have much to add to @leouieda's comment at https://github.com/pyOpenSci/software-review/issues/43#issuecomment-1007459775, but I do use notebooks a lot myself and usually use [Jupytext](https://jupytext.readthedocs.io/en/latest) to keep a synced .ipynb and .py file while doing exploratory data analysis. Do check out this blog post https://devblog.pytorchlightning.ai/publishing-lightning-tutorials-cbea3eaa4b2c which mentions a similar pipeline on turning Python scripts into rendered Jupyter notebook files (not what PyGMT uses, but the core concept of keeping things lightweight is similar).

> @SimonMolinsky - 2. There are cryptic errors which are coming from **GMT**. The example - I'll change borders of the region bounding box that lower coordinate comes after the higher:
>
> I understand that it is a lot of work to catch those exceptions and make them more readable. At the same time I highly recommend to do so. From my perspective, errors and exceptions from the wrapped C/C++ libraries sometimes may generate more noise than help, and it may affect the decision if package will be used or not. (The good example are GDAL errors, or TensorFlow errors - we are not happy if we need to set software based on those packages in production).

A little similar to the first comment above, but on GMT error messages. Again, a long term issue (see e.g. https://github.com/GenericMappingTools/pygmt/issues/256) which has its own set of technical challenges. Off the top of my head, we'll need to either 1) parse the user's inputs and make sure it is correct; and/or 2) parse the GMT error message and return a more user-friendly error message to the user. Both of these impose extra computational costs on the user (in terms of extra if-then checks), but then again, speed-orientated users would probably go with pure GMT anyway. That, and the fact that GMT has thousands of possible error messages from its accumulated three decades of development will make this a tricky beast to handle :slightly_smiling_face:

Anyways, thanks again to both reviewers and @lwasser for your time! Give us a couple of weeks to work out the issues highlighted in https://github.com/pyOpenSci/software-review/issues/43#issuecomment-1007875364 and I'll give you all a ping once those tasks are done.
