### 2022-01-12 #43 @SimonMolinsky
Hi @leouieda , Hi @weiji14

I suspected that tutorials written as Python scripts must make some sense but I didn't consider their memory size. It's invaluable knowledge for me too! I'll consider the same action within my projects.

I must say, that your package is like a spider's web. I've used it once when I was checking its capabilities and I will gradually move all my projects into it. It saves me a lot of time with GUI apps :) I hope that pyGTM will gain a lot of deserved attention.
