### 2022-01-20 GenericMappingTools/pygmt#1686 @meghanrjones
Regarding the CI issue raised in https://github.com/GenericMappingTools/pygmt/issues/1666#issuecomment-993962742, here are two options:

1. Use the [doctest skip directive](https://docs.pytest.org/en/latest/how-to/doctest.html#skipping-tests) to skip particular examples all the time.
2. Add a new target to the Makefile to [invoke pytest without doctest](https://docs.pytest.org/en/6.2.x/usage.html#disabling-plugins), such that the tests without docstring examples can be run on every PR and doctests could be run occasionally.
