### 2022-01-28 GenericMappingTools/pygmt#1686 @meghanrjones
> Should this only be for data processing modules that can't be as easily demonstrated as a plotting function, or should there be in-line examples for all of our modules?

> My vote is for all of them.

Sounds good to me.
