### 2022-01-29 GenericMappingTools/pygmt#1686 @weiji14
> > Should this only be for data processing modules that can't be as easily demonstrated as a plotting function, or should there be in-line examples for all of our modules?
>
> > My vote is for all of them.
>
> Sounds good to me.
>

Ok with doing inline examples for all, but maybe prioritize those that don't already have a gallery example first. E.g. don't do `grdgradient` until much later.
