### 2022-02-13 0404c25
Add inline example for select
Addresses #1686
