### 2022-03-04 GenericMappingTools/pygmt#1697 @meghanrjones
> Does anyone see the same message "This is expected for new tests"? It's not in the PyGMT code.

Wei Ji explained this as a consequence of not running dvc pull before make test during the last community meeting. I recently got this message when I create a new image based test but had not yet added the image to the baseline images directory. The warning is generated by the pytest-mpl plugin.
