### 2022-03-07 GenericMappingTools/pygmt#1686 @weiji14
> 1. Use the [doctest skip directive](https://docs.pytest.org/en/latest/how-to/doctest.html#skipping-tests) to skip particular examples all the time.

Just to cross-reference from #1790, @seisman has found a smart way to skip doctests now using `pytest-doctestplus` and @meghanrjones suggested in https://github.com/GenericMappingTools/pygmt/pull/1790#pullrequestreview-902000046 to document the `__doctest_skip__` and `make fulltest` methods in our contributing guides.

>     2. Add a new target to the Makefile to [invoke pytest without doctest](https://docs.pytest.org/en/6.2.x/usage.html#disabling-plugins), such that the tests without docstring examples can be run on every PR and doctests could be run occasionally.

Just on this, do we want to run `make fulltest` (that includes all of these inline example doctests) on a semi-regular basis? E.g. every week or every month?
