### 2022-03-07 GenericMappingTools/pygmt#1686 @meghanrjones
>
> Just on this, do we want to run `make fulltest` (that includes all of these inline example doctests) on a semi-regular basis? E.g. every week or every month?

I think periodic full tests (e.g., monthly) would be a good idea. A slash command for that workflow could also be helpful for reviewing PRs.
