### 2022-03-08 GenericMappingTools/pygmt#1686 @seisman
I've merged PR #1790 into the main branch. What we still need to do are:

- [ ] Document the `__doctest_skip__ ` and `make fulltest` methods in the contributing guides
- [ ] Run full tests periodically in our CI
- [ ] Implement the slash command workflow (may be optional)
- [ ] Update the existing inline doctests to the new method
