### 2022-03-09 4186400
Wrap grdhisteq (#1433)
Co-authored-by: Wei Ji <23487320+weiji14@users.noreply.github.com>
Co-authored-by: Dongdong Tian <seisman.info@gmail.com>
