### 2022-03-11 GenericMappingTools/pygmt#1665 @seisman
I think this issue actually contains a bug report and a feature request.

### Bug report

In the current implementation, running the following code
```
fig.basemap(region=[-90, -70, 0, 20], projection="M15c")
```
raises an exception `At least one of frame, map_scale, compass, rose, or panel must be specified.`

Because the default value of `frame` is `None`, the following code should raise the same exception.
```
fig.basemap(region=[-90, -70, 0, 20], projection="M15c", frame=None)
```
However, as reported in this issue, the above code doesn't raise the exception. Instead, GMT reports an error `basemap [ERROR]: Must specify at least one of -A, -B, -D, -L, -T`, which make little sense to PyGMT users.

The problem is that when `frame=None` is passed, the `use_alias` decorator converts it to `kwargs["B"]=None`, so the following if-test is `False` and the exception is not raised:
https://github.com/GenericMappingTools/pygmt/blob/418640059915d76f412c2366121aee5e5625e603/pygmt/src/basemap.py#L79-L82

This is a general issue for all aliased options and should be fixed.

### feature request

As discussed in above comments, when none of -A, -B, -D, -L, -T is passed, it makes sense to set `frame=True`, rather than raising an exception.
