### 2022-03-11 d00d347
Rename the parameter 'table' to 'data'
As per https://github.com/GenericMappingTools/pygmt/issues/1479.
