### 2022-03-13 GenericMappingTools/pygmt#1665 @weiji14
Actually, is this a side effect of #1125?
