### 2022-03-14 GenericMappingTools/pygmt#1665 @weiji14
> This is a general issue for all aliased options and should be fixed.

I may have got a generic solution for the bug fix in #1815 by modifying the `args_in_kwargs` function, but needs a bit more testing and handling of edge cases.

> ### feature request
>
> As discussed in above comments, when none of -A, -B, -D, -L, -T is passed, it makes sense to set `frame=True`, rather than raising an exception.

This would need to be handled in a separate Pull Request. I suppose we could modify these lines, so that `kwargs["B"]` is set to `True` if no arguments are passed to A/B/D/L/T.

https://github.com/GenericMappingTools/pygmt/blob/418640059915d76f412c2366121aee5e5625e603/pygmt/src/basemap.py#L79-L82
