### 2022-03-17 GenericMappingTools/pygmt#1665 @seisman
> Actually, is this a side effect of #1125?

Yes, it's related to #1125, but it isn't a side effect of #1125. As I understand the code, the behavior of `frame=None` should not change before or after #1125.

> I may have got a generic solution for the bug fix in #1815 by modifying the `args_in_kwargs` function, but needs a bit more testing and handling of edge cases.
>

I don't think your solution in #1815 is general, because sometimes we don't use `args_in_kwagrs`, instead, we simply check `if "B" in kwargs` or similar.

I think a more general solution is dealing with `frame=None` (i.e. removing `kwargs["B"]` if it's None) in the `@use_alias` decorator.
