### 2022-03-21 GenericMappingTools/pygmt#1663 @seisman
This issue is labeled "good first issue", but it's unclear to me what's the best way to add the information of mamba? Should we add a new subsection or a short note on the installation page?
