### 2022-03-21 #43 @weiji14
Hi all, apologies for the late reply. We've made a new [PyGMT v0.6.0 release](https://github.com/GenericMappingTools/pygmt/releases/tag/v0.6.0) last week :tada: that has addressed the main items in https://github.com/pyOpenSci/software-review/issues/43#issuecomment-1007875364. You can find the updated documentation page at https://www.pygmt.org/v0.6.0. The key changes relevant to this PyOpenSci review include:

- Updates to the main README.md page (installation instructions, notes on other related projects, plus @meghanrjones has added a new intro video!)
- Including a list of development dependencies that new contributors should install (added in https://github.com/GenericMappingTools/pygmt/pull/1783).
- 14 new inline examples for PyGMT functions (tracked in https://github.com/GenericMappingTools/pygmt/issues/1686, and being worked methodologically by @willschlitzer and @michaelgrund)

@lwasser, please let us know what else we'll need to do to complete the review process. Thanks again for your time :smile:
