### 2022-03-28 #43 @lwasser
Hi there! Just a quick reply w an update.. I'm going to try to look at
this..this week. I'm taking a little vacation work break while pyopensci
gets transferred to a new home so it can be it's own organization! I just
was evacuated this weekend because of a fire that's still burning by my
home. So while I'm ok I'm just a bit behind and not focused this week.. I
will check back in when I'm more focused so we can keep this review moving
forward. I truly appreciate the patience. Once I'm working full time again
pyopensci will move forward more efficiently!
