### 2022-03-28 GenericMappingTools/pygmt#1665 @meghanrjones
> Agree with using .get, but we'll need to make sure it treats 0 as a number and not as a boolean False (see https://github.com/GenericMappingTools/pygmt/pull/1815#discussion_r825541819).

This isn't a solution for https://github.com/GenericMappingTools/pygmt/pull/1815#discussion_r825541819, but elsewhere ``if kwargs.get(key) is None:`` should not be affected by this edge-case.
