### 2022-03-29 GenericMappingTools/pygmt#1665 @seisman
> > > I may have got a generic solution for the bug fix in #1815 by modifying the `args_in_kwargs` function, but needs a bit more testing and handling of edge cases.
> >
> >
> > I don't think your solution in #1815 is general, because sometimes we don't use `args_in_kwagrs`, instead, we simply check `if "B" in kwargs` or similar.
> > I think a more general solution is dealing with `frame=None` (i.e. removing `kwargs["B"]` if it's None) in the `@use_alias` decorator.
>
> I think we should refactor those cases that use logic like `if "B" in kwargs"` to use `get` rather than adding more functionality to the `use_alias` decorator.

I'm OK with the `.get` method, but we must be very careful not to use `if "B" in kwargs` when writing new codes.
