### 2022-03-30 GenericMappingTools/pygmt#1665 @seisman
@weiji14 Do you plan to do the `.get` refactor in #1815? Or someone does it in a separate PR?
