### 2022-03-30 GenericMappingTools/pygmt#1665 @weiji14
Yep, I can refactor #1815 to put the logic in the `@use_alias` decorator. Just give me a few hours...
