### 2022-03-30 GenericMappingTools/pygmt#1665 @seisman
> Yep, I can refactor #1815 to put the logic in the `@use_alias` decorator. Just give me a few hours...

Wait... I thought in https://github.com/GenericMappingTools/pygmt/issues/1665#issuecomment-1081039054 @meghanrjones meant to not change the `@use_alias` decorator.
