### 2022-03-30 GenericMappingTools/pygmt#1665 @meghanrjones
> Oh right, so we're using .get() on a case by case basis then? I can go with that.

Yes, I prefer case-by-base because I think the `@use_alias` decorator already does too much (i.e., is too complicated and not very readable) and that we should look towards getting rid of it once all aliases are wrapped so that the code in the modules is more readable (e.g., ``kwargs.get("frame")`` versus ``kwargs.get("B")``) (xref https://github.com/GenericMappingTools/pygmt/issues/262, which originally referenced getting rid of ``@use_alias``).

Fun opportunity to use walrus operators now that we dropped 3.7, I am a bit jealous 😄
