### 2022-03-30 GenericMappingTools/pygmt#1665 @weiji14
Actually, I might do 2 PRs
1. Refactor `args_in_kwargs` in #1815 which should fix the problem in `basemap` and other functions that use `args_in_kwargs`.
2. Replace `if "I" in kwargs` with `kwargs.get("I") is not None` to fix #1852 and other functions that do that.
