### 2022-04-03 GenericMappingTools/pygmt#1665 @seisman
> As discussed in above comments, when none of -A, -B, -D, -L, -T is passed, it makes sense to set `frame=True`, rather than raising an exception.

I'm reopening the issue to track the above feature request.
