### 2022-05-05 GenericMappingTools/pygmt#1686 @willschlitzer
Looking through the examples in the GMT documentation and PyGMT tests, a few of them call remote datasets without using a PyGMT (e.g. `@Table_5_11_mean.xyz` for `surface`). Should the example just call the remote dataset as in, or should `load_sample_data` be updated to include the remote datasets for the in-line example?
