### 2022-05-05 GenericMappingTools/pygmt#1686 @meghanrjones
> Looking through the examples in the GMT documentation and PyGMT tests, a few of them call remote datasets without using a PyGMT function (e.g. @Table_5_11_mean.xyz for surface). Should the example just call the remote dataset as is (with the "@" in front calling the file), or should load_sample_data() be updated to include the remote datasets for the in-line example?

For tests, IMO it doesn't matter much. For user-facing examples I think it would be better to update `load_sample_data`.
