### 2022-05-06 GenericMappingTools/pygmt#1686 @willschlitzer
Are there any gridded datasets with holes in them to use for an example for `grdfill`. I can use the code from one of the tests, but it seems counterproductive to demo removing data from a grd file only to add it back in with `grdfill`.
