### 2022-05-06 GenericMappingTools/pygmt#1686 @willschlitzer
> > Are there any gridded datasets with holes in them to use for an example for grdfill? I can use the code from one of the tests, but it seems counterproductive to demo removing data from a grd file only to add it back in with grdfill.
>
> Looking through https://oceania.generic-mapping-tools.org/cache/, `data_w_nans.nc` seems like one option. Below is a picture from the gmt tests that uses that file.
>
> <img alt="image" width="314" src="https://user-images.githubusercontent.com/14077947/167139243-1d689839-3bc1-4595-acb6-401dc1acaf3c.png">

Thanks @meghanrjones!
