### 2022-05-06 GenericMappingTools/pygmt#1686 @meghanrjones
> Are there any gridded datasets with holes in them to use for an example for `grdfill`? I can use the code from one of the tests, but it seems counterproductive to demo removing data from a grd file only to add it back in with `grdfill`.

There's now a better file to use - `earth_relief_20m_holes.grd`, see https://github.com/GenericMappingTools/gmt/pull/6678 and https://github.com/GenericMappingTools/gmtserver-admin/pull/155 for background.
