### 2022-06-28 #43 @lwasser
hey there @weiji14 i just wanted to check in as i know i've been stereo silent for too long here. i am currently moving this project to a new home and taking time off. i'm going to check in with the reviewers here to see if the changes made to pygmt satisfy their reviews... i am so sorry that this submission has gotten stuck in my job transition . Please know that in the fall (in about 2 months) i will be back here working full time on this project so this lack of progress on this review doesn't reflect the organization it just reflects this change i'm making in where the project lives and where i work!
