### 2022-06-29 #43 @jbusecke
Hi @lwasser, I anticipate a larger than usual workload in anticipation of Scipy2022. Would a response after July 18th be acceptable? Very sorry to slow down the review.
