### 2022-06-29 #43 @weiji14
Sure! We'll have a PyGMT v0.7.0 release in early July, and a SciPy 2022 talk by @meghanrjones, so ok with a response around mid to end of July.
