### 2022-07-21 #43 @jbusecke
I finally found some time (and sleep) after Scipy, and have reviewed the changes here.
Big thanks for all the developers for the many changes made in response to my review.
The only thing I noticed was that there are still a few checkmarks missing in https://github.com/GenericMappingTools/pygmt/issues/1686, but given the record of development I have seen here, I am confident these will be addressed soon (so no need to request another review period from my end).
I did check the [packaging guide](https://www.pyopensci.org/contributing-guide/authoring/overview.html) and found that this package exceeds all the requirements.

So from my end this is a go on publication 🚀.

PS: This was such a pleasure to review, I wish science journals would enable a workflow even close to this!
