### 2022-07-21 #43 @SimonMolinsky
Hi @weiji14 , @leouieda , @lwasser , @jbusecke!

I had mostly propositions of improvements (more Python, less GMT), and I've checked all linked resources (projects and discussions) - I think that we are on the same page, and this is what I was looking for. I agree, that classes could be better data structures to pass "Pythonic" parameters into drawing functions (instead of dictionaries).

Your new training materials from [here](https://www.generic-mapping-tools.org/egu22pygmt/intro.html) are excellent, and it is good to see that the package is developed as a project with the future and you build a healthy ecosystem around it.

I've updated my review points too, and from my perspective, you are ready for the next publication steps! Thank you for your tremendous work and a lot of effort put into the promotion :)
