### 2022-09-01 #43 @lwasser
@jbusecke @SimonMolinsky wonderful - thank you both SO VERY MUCH for your thoughtful replies. And MY APOLOGIES for how delayed my response was. I've been offline but am not working full time on this project so you can expect future review processes to move forward much more smoothly. Based upon this all, @weiji14 @leouieda I wanted to congratulate you on being accepted by the pyOpenSci review process. And here comes out stock congratulations + next steps:

@weiji14 i will update below once you let me know if you plan to submit to JOSS as well. Once you've submitted to use your JOSS acceptance is fast tracked. This will give you a cross-ref accessible DOI which is nice - you don't have to go through another review. Please just let me know if you wish to do this. If not, please follow the steps below and i'll check of and close this issue once we see the badge on your readme, etc. Get in touch with any questions!!
