### 2022-09-01 #43 @lwasser
----------------------------------------------
🎉 pyGMT has been approved by pyOpenSci! Thank you @weiji14  for submitting pyGMT and many thanks to @SimonMolinsky  @jbusecke for reviewing this package! 😸

- [ ] Activate Zenodo watching the repo if you haven't already done so.
- [ ] Add the badge for pyOpenSci peer-review to the README.md of pyGMT. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`
- [ ] Add <package-name> to the pyOpenSci website. @weiji14 , please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors
- [ ] <reviewers-and-maintainers> if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!

<IF JOSS SUBMISSION>

- [ ] Wait for a JOSS editor to approve the presubmission (which includes a scope check)
- [ ] When the JOSS review is complete, add a comment to your review in the pyOpenSci software-review repo that it has been approved by JOSS.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉
<IF JOSS SUBMISSION/>

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
