### 2022-09-06 #43 @weiji14
Thanks @lwasser and all the reviewers! Just want to quickly acknowledge that we're delighted with the news, and it's been a pleasure to have such a nice open review process :smiley:

In terms of next steps, we've got Zenodo set-up already at https://doi.org/10.5281/zenodo.3781524 (since v0.1.0 in fact), so you can tick the first 2 boxes. I'll submit PRs for the other 3 checkboxes soon (once I get some free capacity).

For the JOSS submission part, let me get back to you on that once the team has a chat on https://github.com/GenericMappingTools/pygmt/issues/677#issuecomment-1238211476. Just to understand the expedited process though (as I'm a little new to JOSS), would the JOSS review just be focused on the [software paper](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) itself and not the code/documentation? I had a look at https://www.pyopensci.org/contributing-guide/open-source-software-submissions/author-guide.html#journal-of-open-source-software-joss-submission which mentions writing a `paper.md` file, but not sure if there's anything else needed.
