### 2022-09-06 #43 @weiji14
> @weiji14 are you on twitter? i'm going to post tomorrow about this review being accepted and would love to mention you.

Yes, I'm on Twitter, but you can just tag [`@gmt_dev`](https://twitter.com/gmt_dev)/#PyGMT and that'd be great. Prefer to give credit to the whole PyGMT team in the spirit of open source :grin:
