### 2022-09-06 #43 @lwasser
> Yes, I'm on Twitter, but you can just tag [`@gmt_dev`](https://twitter.com/gmt_dev)/#PyGMT and that'd be great. Prefer to give credit to the whole PyGMT team in the spirit of open source 😁

ok wonderful :) thanks!
