### 2022-09-06 #43 @jbusecke
Wonderful news! Thanks for letting me be part of this (by miles best I have ever had) reviewer experience. Also very delighted to learn about the overlap with JOSS.
