### 2022-09-07 #43 @lwasser
@jbusecke thank you!! and that's really wonderful to hear that you thought the process was good as well. In the future the process won't be as delayed (i've been offline for the past 6 months so that was my fault!).

The last step for reviewers is adding @SimonMolinsky and @jbusecke to[ our website as contributors, here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) .  Are you comfortable with adding yourselves via a PR to that repo and that file?

Many thanks!
