### 2022-09-07 9e2e350
Add PyOpenSci peer reviewed badge to main README.rst
Pointing to https://github.com/pyOpenSci/software-review/issues/43.
