### 2022-09-08 cbe3a5d
Add PyOpenSci peer reviewed badge to main README.rst (#2112)
Pointing to https://github.com/pyOpenSci/software-review/issues/43.
