### 2022-09-29 #43 @SimonMolinsky
> hey there @weiji14 @jbusecke @SimonMolinsky 👋 Just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it. Thank you in advance for helping us by filling out the survey!! 🙌
>
> Wei, it's really important for us to collect information from our maintainers so that we can both stay in touch with you regarding package maintenance and also support you through time. We really appreciate your time in filling this out. I know that you have a team working on pyGMT. Please have your co-maintainers also fill it out and please list them here as well. 🙌 Many thanks in advance! 🙌
>
> [🔗 HERE IS THE SURVEY LINK 🔗](https://forms.gle/ZZXcD1hr18tYmY2p6)

Done! 😊
