### 2023-06-16 #43 @juanis2112
I'll open an issue to start a discussion with the ideas people have posted and maybe do some voting to decide which direction to explore more.
