### 2023-08-24 #43 @weiji14
Hi @lwasser, sorry for the late reply! Our list of maintainers are listed at https://www.pygmt.org/v0.9.0/team.html:

- `@seisman`
- `@maxrjones`
- `@michaelgrund`
- `@weiji14`
- `@willschlitzer`

Could you post the link on the website when this is updated? Thanks!
