### 2023-09-11 #43 @lwasser
@weiji14 no worries! thank you for replying! i just wanted to be sure we are recognizing all of the maintainers on our [package listing page](https://www.pyopensci.org/python-packages.html). i've added everyone to this issue so the data will get updated in a few weeks (i have a bug in our code to fix and then it can be updated!!). i hope all is well!!

once i perform that update (i could do it now if that bug didn't exist) i'll let you know and add a screen shot here! but i did update the issue which the code parses so everyone gets recognized!!
