### 2023-09-11 #43 @lwasser
@weiji14 i just updated the website!! all maintainers are now credited on the pygmt package page!! And they should be listed on our contributor page too!!

in the future you will also be able to filter by maintainer on our [contributor page](https://www.pyopensci.org/our-community/index.html#pyopensci-community-contributors) - that is a feature i plan to add!

thank you for helping me flesh out the maintainer list here. we really want to make sure all of the authors / maintainers get credit for the great work done on these tools!
