### 2023-09-13 #43 @weiji14
Cool, I've updated the maintainer list above, and see that the [cron-job runs on the 1st and 15th of each month](https://github.com/pyOpenSci/pyopensci.github.io/blob/b9bdaf3f1d951c715c0d8887d417c42e9e66e3f9/.github/workflows/update-contribs-reviews.yml#L7), which is tomorrow, so should be ok to wait for a bit. Nice work on automating all of this!
