### 2023-09-13 #43 @lwasser
Yes! you got it! so it should open a pr and then everything will get updated.

i forget if we already talked about this - but do you (or anyone else involved in this review including other maintainers!!) have any interest in joining our slack? if so i can send invites to anyone who's interested.
