### 2023-09-14 #43 @weiji14
I'll personally pass for the Slack channel (have more than a dozen channels already...), but maybe someone else would be interested.
